% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_alias}
\alias{gamelift_create_alias}
\title{Creates an alias for a fleet}
\usage{
gamelift_create_alias(Name, Description, RoutingStrategy)
}
\arguments{
\item{Name}{[required] Descriptive label that is associated with an alias. Alias names do not
need to be unique.}

\item{Description}{Human-readable description of an alias.}

\item{RoutingStrategy}{[required] Object that specifies the fleet and routing type to use for the alias.}
}
\description{
Creates an alias for a fleet. In most situations, you can use an alias
ID in place of a fleet ID. By using a fleet alias instead of a specific
fleet ID, you can switch gameplay and players to a new fleet without
changing your game client or other game components. For example, for
games in production, using an alias allows you to seamlessly redirect
your player base to a new game server update.
}
\details{
Amazon GameLift supports two types of routing strategies for aliases:
simple and terminal. A simple alias points to an active fleet. A
terminal alias is used to display messaging or link to a URL instead of
routing players to an active fleet. For example, you might use a
terminal alias when a game version is no longer supported and you want
to direct players to an upgrade site.

To create a fleet alias, specify an alias name, routing strategy, and
optional description. Each simple alias can point to only one fleet, but
a fleet can have multiple aliases. If successful, a new alias record is
returned, including an alias ID, which you can reference when creating a
game session. You can reassign an alias to another fleet by calling
\code{UpdateAlias}.
\itemize{
\item CreateAlias
\item ListAliases
\item DescribeAlias
\item UpdateAlias
\item DeleteAlias
\item ResolveAlias
}
}
\section{Request syntax}{
\preformatted{svc$create_alias(
  Name = "string",
  Description = "string",
  RoutingStrategy = list(
    Type = "SIMPLE"|"TERMINAL",
    FleetId = "string",
    Message = "string"
  )
)
}
}

\keyword{internal}
