% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_fleet_locations}
\alias{gamelift_create_fleet_locations}
\title{Adds remote locations to a fleet and begins populating the new locations
with EC2 instances}
\usage{
gamelift_create_fleet_locations(FleetId, Locations)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to add locations to. You can use
either the fleet ID or ARN value.}

\item{Locations}{[required] A list of locations to deploy additional instances to and manage as part
of the fleet. You can add any GameLift-supported AWS Region as a remote
location, in the form of an AWS Region code such as \code{us-west-2}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetId = "string",
  FleetArn = "string",
  LocationStates = list(
    list(
      Location = "string",
      Status = "NEW"|"DOWNLOADING"|"VALIDATING"|"BUILDING"|"ACTIVATING"|"ACTIVE"|"DELETING"|"ERROR"|"TERMINATED"
    )
  )
)
}
}
\description{
Adds remote locations to a fleet and begins populating the new locations
with EC2 instances. The new instances conform to the fleet's instance
type, auto-scaling, and other configuration settings.

This operation cannot be used with fleets that don't support remote
locations. Fleets can have multiple locations only if they reside in AWS
Regions that support this feature (see
\code{\link[=gamelift_create_fleet]{create_fleet}} for the complete list) and were
created after the feature was released in March 2021.

To add fleet locations, specify the fleet to be updated and provide a
list of one or more locations.

If successful, this operation returns the list of added locations with
their status set to \code{NEW}. GameLift initiates the process of starting an
instance in each added location. You can track the status of each new
location by monitoring location creation events using
\code{\link[=gamelift_describe_fleet_events]{describe_fleet_events}}.
Alternatively, you can poll location status by calling
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}.
After a location status becomes \code{ACTIVE}, you can adjust the location's
capacity as needed with
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}}.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Multi-location fleets}

\strong{Related actions}

\code{\link[=gamelift_create_fleet_locations]{create_fleet_locations}} |
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}
|
\code{\link[=gamelift_describe_fleet_location_capacity]{describe_fleet_location_capacity}}
|
\code{\link[=gamelift_describe_fleet_location_utilization]{describe_fleet_location_utilization}}
| \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}} |
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}} |
\code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}} |
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}} |
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}} |
\code{\link[=gamelift_delete_fleet_locations]{delete_fleet_locations}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$create_fleet_locations(
  FleetId = "string",
  Locations = list(
    list(
      Location = "string"
    )
  )
)
}
}

\keyword{internal}
