% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_game_session}
\alias{gamelift_create_game_session}
\title{Creates a multiplayer game session for players in a specific fleet
location}
\usage{
gamelift_create_game_session(FleetId, AliasId,
  MaximumPlayerSessionCount, Name, GameProperties, CreatorId,
  GameSessionId, IdempotencyToken, GameSessionData, Location)
}
\arguments{
\item{FleetId}{A unique identifier for the fleet to create a game session in. You can
use either the fleet ID or ARN value. Each request must reference either
a fleet ID or alias ID, but not both.}

\item{AliasId}{A unique identifier for the alias associated with the fleet to create a
game session in. You can use either the alias ID or ARN value. Each
request must reference either a fleet ID or alias ID, but not both.}

\item{MaximumPlayerSessionCount}{[required] The maximum number of players that can be connected simultaneously to
the game session.}

\item{Name}{A descriptive label that is associated with a game session. Session
names do not need to be unique.}

\item{GameProperties}{A set of custom properties for a game session, formatted as key:value
pairs. These properties are passed to a game server process in the
GameSession object with a request to start a new game session.}

\item{CreatorId}{A unique identifier for a player or entity creating the game session.
This parameter is required when requesting a new game session on a fleet
with a resource creation limit policy. This type of policy limits the
number of concurrent active game sessions that one player can create
within a certain time span. GameLift uses the CreatorId to evaluate the
new request against the policy.}

\item{GameSessionId}{\emph{This parameter is no longer preferred. Please use \code{IdempotencyToken}
instead.} Custom string that uniquely identifies a request for a new
game session. Maximum token length is 48 characters. If provided, this
string is included in the new game session's ID.}

\item{IdempotencyToken}{Custom string that uniquely identifies the new game session request.
This is useful for ensuring that game session requests with the same
idempotency token are processed only once. Subsequent requests with the
same string return the original \code{GameSession} object, with an updated
status. Maximum token length is 48 characters. If provided, this string
is included in the new game session's ID. A game session ARN has the
following format:
\verb{arn:aws:gamelift:<region>::gamesession/<fleet ID>/<custom ID string or idempotency token>}.
Idempotency tokens remain in use for 30 days after a game session has
ended; game session objects are retained for this time period and then
deleted.}

\item{GameSessionData}{A set of custom game session properties, formatted as a single string
value. This data is passed to a game server process in the GameSession
object with a request to start a new game session.}

\item{Location}{A fleet's remote location to place the new game session in. If this
parameter is not set, the new game session is placed in the fleet's home
Region. Specify a remote location with an AWS Region code such as
\code{us-west-2}.}
}
\value{
A list with the following syntax:\preformatted{list(
  GameSession = list(
    GameSessionId = "string",
    Name = "string",
    FleetId = "string",
    FleetArn = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    TerminationTime = as.POSIXct(
      "2015-01-01"
    ),
    CurrentPlayerSessionCount = 123,
    MaximumPlayerSessionCount = 123,
    Status = "ACTIVE"|"ACTIVATING"|"TERMINATED"|"TERMINATING"|"ERROR",
    StatusReason = "INTERRUPTED",
    GameProperties = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    IpAddress = "string",
    DnsName = "string",
    Port = 123,
    PlayerSessionCreationPolicy = "ACCEPT_ALL"|"DENY_ALL",
    CreatorId = "string",
    GameSessionData = "string",
    MatchmakerData = "string",
    Location = "string"
  )
)
}
}
\description{
Creates a multiplayer game session for players in a specific fleet
location. This operation prompts an available server process to start a
game session and retrieves connection information for the new game
session. As an alternative, consider using the GameLift game session
placement feature with

with
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}},
which uses FleetIQ algorithms and queues to optimize the placement
process.

When creating a game session, you specify exactly where you want to
place it and provide a set of game session configuration settings. The
fleet must be in \code{ACTIVE} status before a game session can be created in
it.

This operation can be used in the following ways:
\itemize{
\item To create a game session on an instance in a fleet's home Region,
provide a fleet or alias ID along with your game session
configuration.
\item To create a game session on an instance in a fleet's remote
location, provide a fleet or alias ID and a location name, along
with your game session configuration.
}

If successful, a workflow is initiated to start a new game session. A
\code{GameSession} object is returned containing the game session
configuration and status. When the status is \code{ACTIVE}, game session
connection information is provided and player sessions can be created
for the game session. By default, newly created game sessions are open
to new players. You can restrict new player access by using
\code{\link[=gamelift_update_game_session]{update_game_session}} to change the game
session's player session creation policy.

Game session logs are retained for all active game sessions for 14 days.
To access the logs, call
\code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}} to
download the log files.

\emph{Available in GameLift Local.}

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a game session}

\strong{Related actions}

\code{\link[=gamelift_create_game_session]{create_game_session}} |
\code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}} |
\code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
| \code{\link[=gamelift_search_game_sessions]{search_game_sessions}} |
\code{\link[=gamelift_update_game_session]{update_game_session}} |
\code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}} |
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
|
\code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
| \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$create_game_session(
  FleetId = "string",
  AliasId = "string",
  MaximumPlayerSessionCount = 123,
  Name = "string",
  GameProperties = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  CreatorId = "string",
  GameSessionId = "string",
  IdempotencyToken = "string",
  GameSessionData = "string",
  Location = "string"
)
}
}

\keyword{internal}
