% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_matchmaking_configuration}
\alias{gamelift_create_matchmaking_configuration}
\title{Defines a new matchmaking configuration for use with FlexMatch}
\usage{
gamelift_create_matchmaking_configuration(Name, Description,
  GameSessionQueueArns, RequestTimeoutSeconds, AcceptanceTimeoutSeconds,
  AcceptanceRequired, RuleSetName, NotificationTarget,
  AdditionalPlayerCount, CustomEventData, GameProperties, GameSessionData,
  BackfillMode, FlexMatchMode, Tags)
}
\arguments{
\item{Name}{[required] A unique identifier for a matchmaking configuration. This name is used
to identify the configuration associated with a matchmaking request or
ticket.}

\item{Description}{A human-readable description of the matchmaking configuration.}

\item{GameSessionQueueArns}{Amazon Resource Name
(\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN})
that is assigned to a GameLift game session queue resource and uniquely
identifies it. ARNs are unique across all Regions. Queues can be located
in any Region. Queues are used to start new GameLift-hosted game
sessions for matches that are created with this matchmaking
configuration. If \code{FlexMatchMode} is set to \code{STANDALONE}, do not set
this parameter.}

\item{RequestTimeoutSeconds}{[required] The maximum duration, in seconds, that a matchmaking ticket can remain
in process before timing out. Requests that fail due to timing out can
be resubmitted as needed.}

\item{AcceptanceTimeoutSeconds}{The length of time (in seconds) to wait for players to accept a proposed
match, if acceptance is required. If any player rejects the match or
fails to accept before the timeout, the tickets are returned to the
ticket pool and continue to be evaluated for an acceptable match.}

\item{AcceptanceRequired}{[required] A flag that determines whether a match that was created with this
configuration must be accepted by the matched players. To require
acceptance, set to \code{TRUE}. With this option enabled, matchmaking tickets
use the status \code{REQUIRES_ACCEPTANCE} to indicate when a completed
potential match is waiting for player acceptance.}

\item{RuleSetName}{[required] A unique identifier for a matchmaking rule set to use with this
configuration. You can use either the rule set name or ARN value. A
matchmaking configuration can only use rule sets that are defined in the
same Region.}

\item{NotificationTarget}{An SNS topic ARN that is set up to receive matchmaking notifications.}

\item{AdditionalPlayerCount}{The number of player slots in a match to keep open for future players.
For example, assume that the configuration's rule set specifies a match
for a single 12-person team. If the additional player count is set to 2,
only 10 players are initially selected for the match. This parameter is
not used if \code{FlexMatchMode} is set to \code{STANDALONE}.}

\item{CustomEventData}{Information to be added to all events related to this matchmaking
configuration.}

\item{GameProperties}{A set of custom properties for a game session, formatted as key-value
pairs. These properties are passed to a game server process in the
GameSession object with a request to start a new game session (see
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).
This information is added to the new GameSession object that is created
for a successful match. This parameter is not used if \code{FlexMatchMode} is
set to \code{STANDALONE}.}

\item{GameSessionData}{A set of custom game session properties, formatted as a single string
value. This data is passed to a game server process in the GameSession
object with a request to start a new game session (see \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).
This information is added to the new GameSession object that is created
for a successful match. This parameter is not used if \code{FlexMatchMode} is
set to \code{STANDALONE}.}

\item{BackfillMode}{The method used to backfill game sessions that are created with this
matchmaking configuration. Specify \code{MANUAL} when your game manages
backfill requests manually or does not use the match backfill feature.
Specify \code{AUTOMATIC} to have GameLift create a
\code{\link[=gamelift_start_match_backfill]{start_match_backfill}} request whenever
a game session has one or more open slots. Learn more about manual and
automatic backfill in \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-backfill.html}{Backfill Existing Games with FlexMatch}.
Automatic backfill is not available when \code{FlexMatchMode} is set to
\code{STANDALONE}.}

\item{FlexMatchMode}{Indicates whether this matchmaking configuration is being used with
GameLift hosting or as a standalone matchmaking solution.
\itemize{
\item \strong{STANDALONE} - FlexMatch forms matches and returns match
information, including players and team assignments, in a
\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-events.html#match-events-matchmakingsucceeded}{MatchmakingSucceeded}
event.
\item \strong{WITH_QUEUE} - FlexMatch forms matches and uses the specified
GameLift queue to start a game session for the match.
}}

\item{Tags}{A list of labels to assign to the new matchmaking configuration
resource. Tags are developer-defined key-value pairs. Tagging AWS
resources are useful for resource management, access management and cost
allocation. For more information, see \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
in the \emph{AWS General Reference}. Once the resource is created, you can
use \code{\link[=gamelift_tag_resource]{tag_resource}},
\code{\link[=gamelift_untag_resource]{untag_resource}}, and
\code{\link[=gamelift_list_tags_for_resource]{list_tags_for_resource}} to add,
remove, and view tags. The maximum tag limit may be lower than stated.
See the AWS General Reference for actual tagging limits.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Configuration = list(
    Name = "string",
    ConfigurationArn = "string",
    Description = "string",
    GameSessionQueueArns = list(
      "string"
    ),
    RequestTimeoutSeconds = 123,
    AcceptanceTimeoutSeconds = 123,
    AcceptanceRequired = TRUE|FALSE,
    RuleSetName = "string",
    RuleSetArn = "string",
    NotificationTarget = "string",
    AdditionalPlayerCount = 123,
    CustomEventData = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    GameProperties = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    GameSessionData = "string",
    BackfillMode = "AUTOMATIC"|"MANUAL",
    FlexMatchMode = "STANDALONE"|"WITH_QUEUE"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Defines a new matchmaking configuration for use with FlexMatch. Whether
your are using FlexMatch with GameLift hosting or as a standalone
matchmaking service, the matchmaking configuration sets out rules for
matching players and forming teams. If you're also using GameLift
hosting, it defines how to start game sessions for each match. Your
matchmaking system can use multiple configurations to handle different
game scenarios. All matchmaking requests
(\code{\link[=gamelift_start_matchmaking]{start_matchmaking}} or
\code{\link[=gamelift_start_match_backfill]{start_match_backfill}}) identify the
matchmaking configuration to use and provide player attributes
consistent with that configuration.

To create a matchmaking configuration, you must provide the following:
configuration name and FlexMatch mode (with or without GameLift
hosting); a rule set that specifies how to evaluate players and find
acceptable matches; whether player acceptance is required; and the
maximum time allowed for a matchmaking attempt. When using FlexMatch
with GameLift hosting, you also need to identify the game session queue
to use when starting a game session for the match.

In addition, you must set up an Amazon Simple Notification Service (SNS)
to receive matchmaking notifications, and provide the topic ARN in the
matchmaking configuration. An alternative method, continuously polling
ticket status with
\code{\link[=gamelift_describe_matchmaking]{describe_matchmaking}}, is only
suitable for games in development with low matchmaking usage.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/gamelift-match.html}{FlexMatch Developer Guide}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-configuration.html}{Design a FlexMatch Matchmaker}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-notification.html}{Set Up FlexMatch Event Notification}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_matchmaking_configuration]{create_matchmaking_configuration}}
\item \code{\link[=gamelift_describe_matchmaking_configurations]{describe_matchmaking_configurations}}
\item \code{\link[=gamelift_update_matchmaking_configuration]{update_matchmaking_configuration}}
\item \code{\link[=gamelift_delete_matchmaking_configuration]{delete_matchmaking_configuration}}
\item \code{\link[=gamelift_create_matchmaking_rule_set]{create_matchmaking_rule_set}}
\item \code{\link[=gamelift_describe_matchmaking_rule_sets]{describe_matchmaking_rule_sets}}
\item \code{\link[=gamelift_validate_matchmaking_rule_set]{validate_matchmaking_rule_set}}
\item \code{\link[=gamelift_delete_matchmaking_rule_set]{delete_matchmaking_rule_set}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_matchmaking_configuration(
  Name = "string",
  Description = "string",
  GameSessionQueueArns = list(
    "string"
  ),
  RequestTimeoutSeconds = 123,
  AcceptanceTimeoutSeconds = 123,
  AcceptanceRequired = TRUE|FALSE,
  RuleSetName = "string",
  NotificationTarget = "string",
  AdditionalPlayerCount = 123,
  CustomEventData = "string",
  GameProperties = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  GameSessionData = "string",
  BackfillMode = "AUTOMATIC"|"MANUAL",
  FlexMatchMode = "STANDALONE"|"WITH_QUEUE",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
