% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_player_session}
\alias{gamelift_create_player_session}
\title{Reserves an open player slot in a game session for a player}
\usage{
gamelift_create_player_session(GameSessionId, PlayerId, PlayerData)
}
\arguments{
\item{GameSessionId}{[required] A unique identifier for the game session to add a player to.}

\item{PlayerId}{[required] A unique identifier for a player. Player IDs are developer-defined.}

\item{PlayerData}{Developer-defined information related to a player. GameLift does not use
this data, so it can be formatted as needed for use in the game.}
}
\value{
A list with the following syntax:\preformatted{list(
  PlayerSession = list(
    PlayerSessionId = "string",
    PlayerId = "string",
    GameSessionId = "string",
    FleetId = "string",
    FleetArn = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    TerminationTime = as.POSIXct(
      "2015-01-01"
    ),
    Status = "RESERVED"|"ACTIVE"|"COMPLETED"|"TIMEDOUT",
    IpAddress = "string",
    DnsName = "string",
    Port = 123,
    PlayerData = "string"
  )
)
}
}
\description{
Reserves an open player slot in a game session for a player. New player
sessions can be created in any game session with an open slot that is in
\code{ACTIVE} status and has a player creation policy of \code{ACCEPT_ALL}. You
can add a group of players to a game session with
\code{\link[=gamelift_create_player_sessions]{create_player_sessions}}.

To create a player session, specify a game session ID, player ID, and
optionally a set of player data.

If successful, a slot is reserved in the game session for the player and
a new PlayerSession object is returned with a player session ID. The
player references the player session ID when sending a connection
request to the game session, and the game server can use it to validate
the player reservation with the GameLift service. Player sessions cannot
be updated.

\emph{Available in Amazon GameLift Local.}

\strong{Related actions}

\code{\link[=gamelift_create_player_session]{create_player_session}} |
\code{\link[=gamelift_create_player_sessions]{create_player_sessions}} |
\code{\link[=gamelift_describe_player_sessions]{describe_player_sessions}} |
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
|
\code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$create_player_session(
  GameSessionId = "string",
  PlayerId = "string",
  PlayerData = "string"
)
}
}

\keyword{internal}
