% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_player_sessions}
\alias{gamelift_create_player_sessions}
\title{Reserves open slots in a game session for a group of players}
\usage{
gamelift_create_player_sessions(GameSessionId, PlayerIds, PlayerDataMap)
}
\arguments{
\item{GameSessionId}{[required] A unique identifier for the game session to add players to.}

\item{PlayerIds}{[required] List of unique identifiers for the players to be added.}

\item{PlayerDataMap}{Map of string pairs, each specifying a player ID and a set of
developer-defined information related to the player. Amazon GameLift
does not use this data, so it can be formatted as needed for use in the
game. Any player data strings for player IDs that are not included in
the \code{PlayerIds} parameter are ignored.}
}
\value{
A list with the following syntax:\preformatted{list(
  PlayerSessions = list(
    list(
      PlayerSessionId = "string",
      PlayerId = "string",
      GameSessionId = "string",
      FleetId = "string",
      FleetArn = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      TerminationTime = as.POSIXct(
        "2015-01-01"
      ),
      Status = "RESERVED"|"ACTIVE"|"COMPLETED"|"TIMEDOUT",
      IpAddress = "string",
      DnsName = "string",
      Port = 123,
      PlayerData = "string"
    )
  )
)
}
}
\description{
Reserves open slots in a game session for a group of players. New player
sessions can be created in any game session with an open slot that is in
\code{ACTIVE} status and has a player creation policy of \code{ACCEPT_ALL}. To add
a single player to a game session, use
\code{\link[=gamelift_create_player_session]{create_player_session}}.

To create player sessions, specify a game session ID and a list of
player IDs. Optionally, provide a set of player data for each player ID.

If successful, a slot is reserved in the game session for each player,
and new PlayerSession objects are returned with player session IDs. Each
player references their player session ID when sending a connection
request to the game session, and the game server can use it to validate
the player reservation with the GameLift service. Player sessions cannot
be updated.

\emph{Available in Amazon GameLift Local.}

\strong{Related actions}

\code{\link[=gamelift_create_player_session]{create_player_session}} |
\code{\link[=gamelift_create_player_sessions]{create_player_sessions}} |
\code{\link[=gamelift_describe_player_sessions]{describe_player_sessions}} |
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
|
\code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$create_player_sessions(
  GameSessionId = "string",
  PlayerIds = list(
    "string"
  ),
  PlayerDataMap = list(
    "string"
  )
)
}
}

\keyword{internal}
