% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_alias}
\alias{gamelift_delete_alias}
\title{Deletes an alias}
\usage{
gamelift_delete_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] A unique identifier of the alias that you want to delete. You can use
either the alias ID or ARN value.}
}
\value{
An empty list.
}
\description{
Deletes an alias. This operation removes all record of the alias. Game
clients attempting to access a server process using the deleted alias
receive an error. To delete an alias, specify the alias ID to be
deleted.
\itemize{
\item \code{\link[=gamelift_create_alias]{create_alias}}
\item \code{\link[=gamelift_list_aliases]{list_aliases}}
\item \code{\link[=gamelift_describe_alias]{describe_alias}}
\item \code{\link[=gamelift_update_alias]{update_alias}}
\item \code{\link[=gamelift_delete_alias]{delete_alias}}
\item \code{\link[=gamelift_resolve_alias]{resolve_alias}}
}
}
\section{Request syntax}{
\preformatted{svc$delete_alias(
  AliasId = "string"
)
}
}

\keyword{internal}
