% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_build}
\alias{gamelift_delete_build}
\title{Deletes a build}
\usage{
gamelift_delete_build(BuildId)
}
\arguments{
\item{BuildId}{[required] A unique identifier for the build to delete. You can use either the
build ID or ARN value.}
}
\value{
An empty list.
}
\description{
Deletes a build. This operation permanently deletes the build resource
and any uploaded build files. Deleting a build does not affect the
status of any active fleets using the build, but you can no longer
create new fleets with the deleted build.

To delete a build, specify the build ID.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-build-intro.html}{Upload a Custom Server Build}

\strong{Related actions}

\code{\link[=gamelift_create_build]{create_build}} |
\code{\link[=gamelift_list_builds]{list_builds}} |
\code{\link[=gamelift_describe_build]{describe_build}} |
\code{\link[=gamelift_update_build]{update_build}} |
\code{\link[=gamelift_delete_build]{delete_build}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$delete_build(
  BuildId = "string"
)
}
}

\keyword{internal}
