% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_fleet_locations}
\alias{gamelift_delete_fleet_locations}
\title{Removes locations from a multi-location fleet}
\usage{
gamelift_delete_fleet_locations(FleetId, Locations)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to delete locations for. You can use
either the fleet ID or ARN value.}

\item{Locations}{[required] The list of fleet locations to delete. Specify locations in the form of
an AWS Region code, such as \code{us-west-2}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetId = "string",
  FleetArn = "string",
  LocationStates = list(
    list(
      Location = "string",
      Status = "NEW"|"DOWNLOADING"|"VALIDATING"|"BUILDING"|"ACTIVATING"|"ACTIVE"|"DELETING"|"ERROR"|"TERMINATED"
    )
  )
)
}
}
\description{
Removes locations from a multi-location fleet. When deleting a location,
all game server process and all instances that are still active in the
location are shut down.

To delete fleet locations, identify the fleet ID and provide a list of
the locations to be deleted.

If successful, GameLift sets the location status to \code{DELETING}, and
begins to shut down existing server processes and terminate instances in
each location being deleted. When completed, the location status changes
to \code{TERMINATED}.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift fleets}

\strong{Related actions}

\code{\link[=gamelift_create_fleet_locations]{create_fleet_locations}} |
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}
|
\code{\link[=gamelift_describe_fleet_location_capacity]{describe_fleet_location_capacity}}
|
\code{\link[=gamelift_describe_fleet_location_utilization]{describe_fleet_location_utilization}}
| \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}} |
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}} |
\code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}} |
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}} |
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}} |
\code{\link[=gamelift_delete_fleet_locations]{delete_fleet_locations}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$delete_fleet_locations(
  FleetId = "string",
  Locations = list(
    "string"
  )
)
}
}

\keyword{internal}
