% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_game_server_group}
\alias{gamelift_delete_game_server_group}
\title{This action is part of Amazon GameLift FleetIQ with game server groups,
which is in preview release and is subject to change}
\usage{
gamelift_delete_game_server_group(GameServerGroupName, DeleteOption)
}
\arguments{
\item{GameServerGroupName}{[required] The unique identifier of the game server group to delete. Use either the
GameServerGroup name or ARN value.}

\item{DeleteOption}{The type of delete to perform. Options include:
\itemize{
\item SAFE\\_DELETE – Terminates the game server group and EC2 Auto Scaling
group only when it has no game servers that are in IN\\_USE status.
\item FORCE\\_DELETE – Terminates the game server group, including all
active game servers regardless of their utilization status, and the
EC2 Auto Scaling group.
\item RETAIN – Does a safe delete of the game server group but retains the
EC2 Auto Scaling group as is.
}}
}
\description{
\strong{This action is part of Amazon GameLift FleetIQ with game server
groups, which is in preview release and is subject to change.}
}
\details{
Terminates a game server group and permanently deletes the game server
group record. You have several options for how these resources are
impacted when deleting the game server group. Depending on the type of
delete action selected, this action may affect three types of resources:
the game server group, the corresponding Auto Scaling group, and all
game servers currently running in the group.

To delete a game server group, identify the game server group to delete
and specify the type of delete action to initiate. Game server groups
can only be deleted if they are in ACTIVE or ERROR status.

If the delete request is successful, a series of actions are kicked off.
The game server group status is changed to DELETE\\_SCHEDULED, which
prevents new game servers from being registered and stops autoscaling
activity. Once all game servers in the game server group are
de-registered, GameLift FleetIQ can begin deleting resources. If any of
the delete actions fail, the game server group is placed in ERROR
status.

GameLift FleetIQ emits delete events to Amazon CloudWatch.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item CreateGameServerGroup
\item ListGameServerGroups
\item DescribeGameServerGroup
\item UpdateGameServerGroup
\item DeleteGameServerGroup
\item ResumeGameServerGroup
\item SuspendGameServerGroup
}
}
\section{Request syntax}{
\preformatted{svc$delete_game_server_group(
  GameServerGroupName = "string",
  DeleteOption = "SAFE_DELETE"|"FORCE_DELETE"|"RETAIN"
)
}
}

\keyword{internal}
