% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_game_session_queue}
\alias{gamelift_delete_game_session_queue}
\title{Deletes a game session queue}
\usage{
gamelift_delete_game_session_queue(Name)
}
\arguments{
\item{Name}{[required] A descriptive label that is associated with game session queue. Queue
names must be unique within each Region. You can use either the queue ID
or ARN value.}
}
\value{
An empty list.
}
\description{
Deletes a game session queue. Once a queue is successfully deleted,
unfulfilled
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
requests that reference the queue will fail. To delete a queue, specify
the queue name.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/queues-intro.html}{Using Multi-Region Queues}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_game_session_queue]{create_game_session_queue}}
\item \code{\link[=gamelift_describe_game_session_queues]{describe_game_session_queues}}
\item \code{\link[=gamelift_update_game_session_queue]{update_game_session_queue}}
\item \code{\link[=gamelift_delete_game_session_queue]{delete_game_session_queue}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_game_session_queue(
  Name = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
