% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_matchmaking_rule_set}
\alias{gamelift_delete_matchmaking_rule_set}
\title{Deletes an existing matchmaking rule set}
\usage{
gamelift_delete_matchmaking_rule_set(Name)
}
\arguments{
\item{Name}{[required] A unique identifier for a matchmaking rule set to be deleted. (Note: The
rule set name is different from the optional "name" field in the rule
set body.) You can use either the rule set name or ARN value.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Deletes an existing matchmaking rule set. To delete the rule set,
provide the rule set name. Rule sets cannot be deleted if they are
currently being used by a matchmaking configuration.

\strong{Learn more}
\itemize{
\item \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-rulesets.html}{Build a Rule Set}
}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_matchmaking_configuration]{create_matchmaking_configuration}}
\item \code{\link[=gamelift_describe_matchmaking_configurations]{describe_matchmaking_configurations}}
\item \code{\link[=gamelift_update_matchmaking_configuration]{update_matchmaking_configuration}}
\item \code{\link[=gamelift_delete_matchmaking_configuration]{delete_matchmaking_configuration}}
\item \code{\link[=gamelift_create_matchmaking_rule_set]{create_matchmaking_rule_set}}
\item \code{\link[=gamelift_describe_matchmaking_rule_sets]{describe_matchmaking_rule_sets}}
\item \code{\link[=gamelift_validate_matchmaking_rule_set]{validate_matchmaking_rule_set}}
\item \code{\link[=gamelift_delete_matchmaking_rule_set]{delete_matchmaking_rule_set}}
}
}
\section{Request syntax}{
\preformatted{svc$delete_matchmaking_rule_set(
  Name = "string"
)
}
}

\keyword{internal}
