% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_vpc_peering_connection}
\alias{gamelift_delete_vpc_peering_connection}
\title{Removes a VPC peering connection}
\usage{
gamelift_delete_vpc_peering_connection(FleetId, VpcPeeringConnectionId)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet. This fleet specified must match the
fleet referenced in the VPC peering connection record. You can use
either the fleet ID or ARN value.}

\item{VpcPeeringConnectionId}{[required] A unique identifier for a VPC peering connection. This value is included
in the VpcPeeringConnection object, which can be retrieved by calling
\code{\link[=gamelift_describe_vpc_peering_connections]{describe_vpc_peering_connections}}.}
}
\value{
An empty list.
}
\description{
Removes a VPC peering connection. To delete the connection, you must
have a valid authorization for the VPC peering connection that you want
to delete. You can check for an authorization by calling
\code{\link[=gamelift_describe_vpc_peering_authorizations]{describe_vpc_peering_authorizations}}
or request a new one using
\code{\link[=gamelift_create_vpc_peering_authorization]{create_vpc_peering_authorization}}.

Once a valid authorization exists, call this operation from the AWS
account that is used to manage the Amazon GameLift fleets. Identify the
connection to delete by the connection ID and fleet ID. If successful,
the connection is removed.

\strong{Related actions}

\code{\link[=gamelift_create_vpc_peering_authorization]{create_vpc_peering_authorization}}
|
\code{\link[=gamelift_describe_vpc_peering_authorizations]{describe_vpc_peering_authorizations}}
|
\code{\link[=gamelift_delete_vpc_peering_authorization]{delete_vpc_peering_authorization}}
|
\code{\link[=gamelift_create_vpc_peering_connection]{create_vpc_peering_connection}}
|
\code{\link[=gamelift_describe_vpc_peering_connections]{describe_vpc_peering_connections}}
|
\code{\link[=gamelift_delete_vpc_peering_connection]{delete_vpc_peering_connection}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$delete_vpc_peering_connection(
  FleetId = "string",
  VpcPeeringConnectionId = "string"
)
}
}

\keyword{internal}
