% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_deregister_game_server}
\alias{gamelift_deregister_game_server}
\title{THIS ACTION IS PART OF AMAZON GAMELIFT FLEETIQ WITH GAME SERVER GROUPS,
WHICH IS IN PREVIEW RELEASE AND IS SUBJECT TO CHANGE}
\usage{
gamelift_deregister_game_server(GameServerGroupName, GameServerId)
}
\arguments{
\item{GameServerGroupName}{[required] An identifier for the game server group where the game server to be
de-registered is running. Use either the GameServerGroup name or ARN
value.}

\item{GameServerId}{[required] The identifier for the game server to be de-registered.}
}
\description{
\strong{This action is part of Amazon GameLift FleetIQ with game server
groups, which is in preview release and is subject to change.}
}
\details{
Removes the game server resource from the game server group. As a result
of this action, the de-registered game server can no longer be claimed
and will not returned in a list of active game servers.

To de-register a game server, specify the game server group and game
server ID. If successful, this action emits a CloudWatch event with
termination time stamp and reason.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item RegisterGameServer
\item ListGameServers
\item ClaimGameServer
\item DescribeGameServer
\item UpdateGameServer
\item DeregisterGameServer
}
}
\section{Request syntax}{
\preformatted{svc$deregister_game_server(
  GameServerGroupName = "string",
  GameServerId = "string"
)
}
}

\keyword{internal}
