% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_location_attributes}
\alias{gamelift_describe_fleet_location_attributes}
\title{Retrieves information on a fleet's remote locations, including
life-cycle status and any suspended fleet activity}
\usage{
gamelift_describe_fleet_location_attributes(FleetId, Locations, Limit,
  NextToken)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to retrieve remote locations for. You
can use either the fleet ID or ARN value.}

\item{Locations}{A list of fleet locations to retrieve information for. Specify locations
in the form of an AWS Region code, such as \code{us-west-2}.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages. This limit is
not currently enforced.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetId = "string",
  FleetArn = "string",
  LocationAttributes = list(
    list(
      LocationState = list(
        Location = "string",
        Status = "NEW"|"DOWNLOADING"|"VALIDATING"|"BUILDING"|"ACTIVATING"|"ACTIVE"|"DELETING"|"ERROR"|"TERMINATED"
      ),
      StoppedActions = list(
        "AUTO_SCALING"
      ),
      UpdateStatus = "PENDING_UPDATE"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves information on a fleet's remote locations, including
life-cycle status and any suspended fleet activity.

This operation can be used in the following ways:
\itemize{
\item To get data for specific locations, provide a fleet identifier and a
list of locations. Location data is returned in the order that it is
requested.
\item To get data for all locations, provide a fleet identifier only.
Location data is returned in no particular order.
}

When requesting attributes for multiple locations, use the pagination
parameters to retrieve results as a set of sequential pages.

If successful, a \code{LocationAttributes} object is returned for each
requested location. If the fleet does not have a requested location, no
information is returned. This operation does not return the home Region.
To get information on a fleet's home Region, call
\code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}}.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift fleets}

\strong{Related actions}

\code{\link[=gamelift_create_fleet_locations]{create_fleet_locations}} |
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}
|
\code{\link[=gamelift_describe_fleet_location_capacity]{describe_fleet_location_capacity}}
|
\code{\link[=gamelift_describe_fleet_location_utilization]{describe_fleet_location_utilization}}
| \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}} |
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}} |
\code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}} |
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}} |
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}} |
\code{\link[=gamelift_delete_fleet_locations]{delete_fleet_locations}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_location_attributes(
  FleetId = "string",
  Locations = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
