% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_location_capacity}
\alias{gamelift_describe_fleet_location_capacity}
\title{Retrieves the resource capacity settings for a fleet location}
\usage{
gamelift_describe_fleet_location_capacity(FleetId, Location)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to request location capacity for. You
can use either the fleet ID or ARN value.}

\item{Location}{[required] The fleet location to retrieve capacity information for. Specify a
location in the form of an AWS Region code, such as \code{us-west-2}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetCapacity = list(
    FleetId = "string",
    FleetArn = "string",
    InstanceType = "t2.micro"|"t2.small"|"t2.medium"|"t2.large"|"c3.large"|"c3.xlarge"|"c3.2xlarge"|"c3.4xlarge"|"c3.8xlarge"|"c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"c5a.large"|"c5a.xlarge"|"c5a.2xlarge"|"c5a.4xlarge"|"c5a.8xlarge"|"c5a.12xlarge"|"c5a.16xlarge"|"c5a.24xlarge"|"r3.large"|"r3.xlarge"|"r3.2xlarge"|"r3.4xlarge"|"r3.8xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"r5a.large"|"r5a.xlarge"|"r5a.2xlarge"|"r5a.4xlarge"|"r5a.8xlarge"|"r5a.12xlarge"|"r5a.16xlarge"|"r5a.24xlarge"|"m3.medium"|"m3.large"|"m3.xlarge"|"m3.2xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge"|"m5a.large"|"m5a.xlarge"|"m5a.2xlarge"|"m5a.4xlarge"|"m5a.8xlarge"|"m5a.12xlarge"|"m5a.16xlarge"|"m5a.24xlarge",
    InstanceCounts = list(
      DESIRED = 123,
      MINIMUM = 123,
      MAXIMUM = 123,
      PENDING = 123,
      ACTIVE = 123,
      IDLE = 123,
      TERMINATING = 123
    ),
    Location = "string"
  )
)
}
}
\description{
Retrieves the resource capacity settings for a fleet location. The data
returned includes the current capacity (number of EC2 instances) and
some scaling settings for the requested fleet location. Use this
operation to retrieve capacity information for a fleet's remote location
or home Region (you can also retrieve home Region capacity by calling
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}}).

To retrieve capacity data, identify a fleet and location.

If successful, a \code{FleetCapacity} object is returned for the requested
fleet location.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/monitoring-cloudwatch.html#gamelift-metrics-fleet}{GameLift metrics for fleets}

\strong{Related actions}

\code{\link[=gamelift_create_fleet_locations]{create_fleet_locations}} |
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}
|
\code{\link[=gamelift_describe_fleet_location_capacity]{describe_fleet_location_capacity}}
|
\code{\link[=gamelift_describe_fleet_location_utilization]{describe_fleet_location_utilization}}
| \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}} |
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}} |
\code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}} |
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}} |
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}} |
\code{\link[=gamelift_delete_fleet_locations]{delete_fleet_locations}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_location_capacity(
  FleetId = "string",
  Location = "string"
)
}
}

\keyword{internal}
