% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_location_utilization}
\alias{gamelift_describe_fleet_location_utilization}
\title{Retrieves current usage data for a fleet location}
\usage{
gamelift_describe_fleet_location_utilization(FleetId, Location)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to request location utilization for.
You can use either the fleet ID or ARN value.}

\item{Location}{[required] The fleet location to retrieve utilization information for. Specify a
location in the form of an AWS Region code, such as \code{us-west-2}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetUtilization = list(
    FleetId = "string",
    FleetArn = "string",
    ActiveServerProcessCount = 123,
    ActiveGameSessionCount = 123,
    CurrentPlayerSessionCount = 123,
    MaximumPlayerSessionCount = 123,
    Location = "string"
  )
)
}
}
\description{
Retrieves current usage data for a fleet location. Utilization data
provides a snapshot of current game hosting activity at the requested
location. Use this operation to retrieve utilization information for a
fleet's remote location or home Region (you can also retrieve home
Region utilization by calling
\code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}}).

To retrieve utilization data, identify a fleet and location.

If successful, a \code{FleetUtilization} object is returned for the requested
fleet location.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/monitoring-cloudwatch.html#gamelift-metrics-fleet}{GameLift metrics for fleets}

\strong{Related actions}

\code{\link[=gamelift_create_fleet_locations]{create_fleet_locations}} |
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}
|
\code{\link[=gamelift_describe_fleet_location_capacity]{describe_fleet_location_capacity}}
|
\code{\link[=gamelift_describe_fleet_location_utilization]{describe_fleet_location_utilization}}
| \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}} |
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}} |
\code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}} |
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}} |
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}} |
\code{\link[=gamelift_delete_fleet_locations]{delete_fleet_locations}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_location_utilization(
  FleetId = "string",
  Location = "string"
)
}
}

\keyword{internal}
