% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_port_settings}
\alias{gamelift_describe_fleet_port_settings}
\title{Retrieves a fleet's inbound connection permissions}
\usage{
gamelift_describe_fleet_port_settings(FleetId, Location)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to retrieve port settings for. You can
use either the fleet ID or ARN value.}

\item{Location}{A remote location to check for status of port setting updates. Use the
AWS Region code format, such as \code{us-west-2}.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetId = "string",
  FleetArn = "string",
  InboundPermissions = list(
    list(
      FromPort = 123,
      ToPort = 123,
      IpRange = "string",
      Protocol = "TCP"|"UDP"
    )
  ),
  UpdateStatus = "PENDING_UPDATE",
  Location = "string"
)
}
}
\description{
Retrieves a fleet's inbound connection permissions. Connection
permissions specify the range of IP addresses and port settings that
incoming traffic can use to access server processes in the fleet. Game
sessions that are running on instances in the fleet must use connections
that fall in this range.

This operation can be used in the following ways:
\itemize{
\item To retrieve the inbound connection permissions for a fleet, identify
the fleet's unique identifier.
\item To check the status of recent updates to a fleet remote location,
specify the fleet ID and a location. Port setting updates can take
time to propagate across all locations.
}

If successful, a set of IpPermission objects is returned for the
requested fleet ID. When a location is specified, a pending status is
included. If the requested fleet has been deleted, the result set is
empty.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift fleets}

\strong{Related actions}

\code{\link[=gamelift_list_fleets]{list_fleets}} |
\code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
| \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}} |
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}} |
\code{\link[=gamelift_describe_fleet_events]{describe_fleet_events}} |
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}
|
\code{\link[=gamelift_describe_fleet_port_settings]{describe_fleet_port_settings}}
| \code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}} |
\code{\link[=gamelift_describe_runtime_configuration]{describe_runtime_configuration}}
| \code{\link[=gamelift_describe_scaling_policies]{describe_scaling_policies}} |
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_port_settings(
  FleetId = "string",
  Location = "string"
)
}
}

\keyword{internal}
