% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_utilization}
\alias{gamelift_describe_fleet_utilization}
\title{Retrieves utilization statistics for one or more fleets}
\usage{
gamelift_describe_fleet_utilization(FleetIds, Limit, NextToken)
}
\arguments{
\item{FleetIds}{A unique identifier for the fleet(s) to retrieve utilization data for.
You can use either the fleet ID or ARN value. To retrieve attributes for
all current fleets, do not include this parameter.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages. This parameter
is ignored when the request specifies one or a list of fleet IDs.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.
This parameter is ignored when the request specifies one or a list of
fleet IDs.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetUtilization = list(
    list(
      FleetId = "string",
      FleetArn = "string",
      ActiveServerProcessCount = 123,
      ActiveGameSessionCount = 123,
      CurrentPlayerSessionCount = 123,
      MaximumPlayerSessionCount = 123,
      Location = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves utilization statistics for one or more fleets. Utilization
data provides a snapshot of how the fleet's hosting resources are
currently being used. For fleets with remote locations, this operation
retrieves data for the fleet's home Region only. See
\code{\link[=gamelift_describe_fleet_location_utilization]{describe_fleet_location_utilization}}
to get utilization statistics for a fleet's remote locations.

This operation can be used in the following ways:
\itemize{
\item To get utilization data for one or more specific fleets, provide a
list of fleet IDs or fleet ARNs.
\item To get utilization data for all fleets, do not provide a fleet
identifier.
}

When requesting multiple fleets, use the pagination parameters to
retrieve results as a set of sequential pages.

If successful, a FleetUtilization object is returned for each requested
fleet ID, unless the fleet identifier is not found. Each fleet
utilization object includes a \code{Location} property, which is set to the
fleet's home Region.

Some API operations may limit the number of fleet IDs allowed in one
request. If a request exceeds this limit, the request fails and the
error message includes the maximum allowed.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/monitoring-cloudwatch.html#gamelift-metrics-fleet}{GameLift Metrics for Fleets}

\strong{Related actions}

\code{\link[=gamelift_list_fleets]{list_fleets}} |
\code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
| \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}} |
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}} |
\code{\link[=gamelift_describe_fleet_events]{describe_fleet_events}} |
\code{\link[=gamelift_describe_fleet_location_attributes]{describe_fleet_location_attributes}}
|
\code{\link[=gamelift_describe_fleet_port_settings]{describe_fleet_port_settings}}
| \code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}} |
\code{\link[=gamelift_describe_runtime_configuration]{describe_runtime_configuration}}
| \code{\link[=gamelift_describe_scaling_policies]{describe_scaling_policies}} |
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_utilization(
  FleetIds = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
