% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_game_session_queues}
\alias{gamelift_describe_game_session_queues}
\title{Retrieves the properties for one or more game session queues}
\usage{
gamelift_describe_game_session_queues(Names, Limit, NextToken)
}
\arguments{
\item{Names}{A list of queue names to retrieve information for. You can use either
the queue ID or ARN value. To request settings for all queues, leave
this parameter empty.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this action. To
start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves the properties for one or more game session queues. When
requesting multiple queues, use the pagination parameters to retrieve
results as a set of sequential pages. If successful, a GameSessionQueue
object is returned for each requested queue. When specifying a list of
queues, objects are returned only for queues that currently exist in the
Region.
}
\details{
\itemize{
\item CreateGameSessionQueue
\item DescribeGameSessionQueues
\item UpdateGameSessionQueue
\item DeleteGameSessionQueue
}
}
\section{Request syntax}{
\preformatted{svc$describe_game_session_queues(
  Names = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
