% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_player_sessions}
\alias{gamelift_describe_player_sessions}
\title{Retrieves properties for one or more player sessions}
\usage{
gamelift_describe_player_sessions(GameSessionId, PlayerId,
  PlayerSessionId, PlayerSessionStatusFilter, Limit, NextToken)
}
\arguments{
\item{GameSessionId}{A unique identifier for the game session to retrieve player sessions
for.}

\item{PlayerId}{A unique identifier for a player to retrieve player sessions for.}

\item{PlayerSessionId}{A unique identifier for a player session to retrieve.}

\item{PlayerSessionStatusFilter}{Player session status to filter results on.

Possible player session statuses include the following:
\itemize{
\item \strong{RESERVED} -- The player session request has been received, but
the player has not yet connected to the server process and/or been
validated.
\item \strong{ACTIVE} -- The player has been validated by the server process
and is currently connected.
\item \strong{COMPLETED} -- The player connection has been dropped.
\item \strong{TIMEDOUT} -- A player session request was received, but the
player did not connect and/or was not validated within the timeout
limit (60 seconds).
}}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages. If a player
session ID is specified, this parameter is ignored.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value. If
a player session ID is specified, this parameter is ignored.}
}
\value{
A list with the following syntax:\preformatted{list(
  PlayerSessions = list(
    list(
      PlayerSessionId = "string",
      PlayerId = "string",
      GameSessionId = "string",
      FleetId = "string",
      FleetArn = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      TerminationTime = as.POSIXct(
        "2015-01-01"
      ),
      Status = "RESERVED"|"ACTIVE"|"COMPLETED"|"TIMEDOUT",
      IpAddress = "string",
      DnsName = "string",
      Port = 123,
      PlayerData = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves properties for one or more player sessions.

This action can be used in the following ways:
\itemize{
\item To retrieve a specific player session, provide the player session ID
only.
\item To retrieve all player sessions in a game session, provide the game
session ID only.
\item To retrieve all player sessions for a specific player, provide a
player ID only.
}

To request player sessions, specify either a player session ID, game
session ID, or player ID. You can filter this request by player session
status. Use the pagination parameters to retrieve results as a set of
sequential pages.

If successful, a \code{PlayerSession} object is returned for each session
that matches the request.

\emph{Available in Amazon GameLift Local.}

\strong{Related actions}

\code{\link[=gamelift_create_player_session]{create_player_session}} |
\code{\link[=gamelift_create_player_sessions]{create_player_sessions}} |
\code{\link[=gamelift_describe_player_sessions]{describe_player_sessions}} |
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
|
\code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_player_sessions(
  GameSessionId = "string",
  PlayerId = "string",
  PlayerSessionId = "string",
  PlayerSessionStatusFilter = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
