% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_get_game_session_log_url}
\alias{gamelift_get_game_session_log_url}
\title{Retrieves the location of stored game session logs for a specified game
session}
\usage{
gamelift_get_game_session_log_url(GameSessionId)
}
\arguments{
\item{GameSessionId}{[required] A unique identifier for the game session to get logs for.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PreSignedUrl = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves the location of stored game session logs for a specified game
session. When a game session is terminated, Amazon GameLift
automatically stores the logs in Amazon S3 and retains them for 14 days.
Use this URL to download the logs.

See the \href{https://docs.aws.amazon.com/general/latest/gr/aws_service_limits.html#limits_gamelift}{AWS Service Limits}
page for maximum log file sizes. Log files that exceed this limit are
not saved.
\itemize{
\item \code{\link[=gamelift_create_game_session]{create_game_session}}
\item \code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}}
\item \code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
\item \code{\link[=gamelift_search_game_sessions]{search_game_sessions}}
\item \code{\link[=gamelift_update_game_session]{update_game_session}}
\item \code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}}
\item Game session placements
\itemize{
\item \code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
\item \code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
\item \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_game_session_log_url(
  GameSessionId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
