% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_get_instance_access}
\alias{gamelift_get_instance_access}
\title{Requests remote access to a fleet instance}
\usage{
gamelift_get_instance_access(FleetId, InstanceId)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet that contains the instance you want
access to. You can use either the fleet ID or ARN value. The fleet can
be in any of the following statuses: \code{ACTIVATING}, \code{ACTIVE}, or \code{ERROR}.
Fleets with an \code{ERROR} status may be accessible for a short time before
they are deleted.}

\item{InstanceId}{[required] A unique identifier for an instance you want to get access to. You can
access an instance in any status.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  InstanceAccess = list(
    FleetId = "string",
    InstanceId = "string",
    IpAddress = "string",
    OperatingSystem = "WINDOWS_2012"|"AMAZON_LINUX"|"AMAZON_LINUX_2",
    Credentials = list(
      UserName = "string",
      Secret = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Requests remote access to a fleet instance. Remote access is useful for
debugging, gathering benchmarking data, or observing activity in real
time.

To remotely access an instance, you need credentials that match the
operating system of the instance. For a Windows instance, Amazon
GameLift returns a user name and password as strings for use with a
Windows Remote Desktop client. For a Linux instance, Amazon GameLift
returns a user name and RSA private key, also as strings, for use with
an SSH client. The private key must be saved in the proper format to a
\code{.pem} file before using. If you're making this request using the AWS
CLI, saving the secret can be handled as part of the GetInstanceAccess
request, as shown in one of the examples for this operation.

To request access to a specific instance, specify the IDs of both the
instance and the fleet it belongs to. You can retrieve a fleet's
instance IDs by calling
\code{\link[=gamelift_describe_instances]{describe_instances}}. If successful, an
InstanceAccess object is returned that contains the instance's IP
address and a set of credentials.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-remote-access.html}{Remotely Access Fleet Instances}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-creating-debug.html}{Debug Fleet Issues}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_describe_instances]{describe_instances}}
\item \code{\link[=gamelift_get_instance_access]{get_instance_access}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$get_instance_access(
  FleetId = "string",
  InstanceId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
