% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_aliases}
\alias{gamelift_list_aliases}
\title{Retrieves all aliases for this AWS account}
\usage{
gamelift_list_aliases(RoutingStrategyType, Name, Limit, NextToken)
}
\arguments{
\item{RoutingStrategyType}{The routing type to filter results on. Use this parameter to retrieve
only aliases with a certain routing type. To retrieve all aliases, leave
this parameter empty.

Possible routing types include the following:
\itemize{
\item \strong{SIMPLE} -- The alias resolves to one specific fleet. Use this
type when routing to active fleets.
\item \strong{TERMINAL} -- The alias does not resolve to a fleet but instead
can be used to display a message to the user. A terminal alias
throws a TerminalRoutingStrategyException with the RoutingStrategy
message embedded.
}}

\item{Name}{A descriptive label that is associated with an alias. Alias names do not
need to be unique.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\value{
A list with the following syntax:\preformatted{list(
  Aliases = list(
    list(
      AliasId = "string",
      Name = "string",
      AliasArn = "string",
      Description = "string",
      RoutingStrategy = list(
        Type = "SIMPLE"|"TERMINAL",
        FleetId = "string",
        Message = "string"
      ),
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastUpdatedTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves all aliases for this AWS account. You can filter the result
set by alias name and/or routing strategy type. Use the pagination
parameters to retrieve results in sequential pages.

Returned aliases are not listed in any particular order.

\strong{Related actions}

\code{\link[=gamelift_create_alias]{create_alias}} |
\code{\link[=gamelift_list_aliases]{list_aliases}} |
\code{\link[=gamelift_describe_alias]{describe_alias}} |
\code{\link[=gamelift_update_alias]{update_alias}} |
\code{\link[=gamelift_delete_alias]{delete_alias}} |
\code{\link[=gamelift_resolve_alias]{resolve_alias}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$list_aliases(
  RoutingStrategyType = "SIMPLE"|"TERMINAL",
  Name = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
