% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_fleets}
\alias{gamelift_list_fleets}
\title{Retrieves a collection of fleet records for this AWS account}
\usage{
gamelift_list_fleets(BuildId, ScriptId, Limit, NextToken)
}
\arguments{
\item{BuildId}{Unique identifier for a build to return fleets for. Use this parameter
to return only fleets using the specified build. To retrieve all fleets,
leave this parameter empty.}

\item{ScriptId}{Unique identifier for a Realtime script to return fleets for. Use this
parameter to return only fleets using the specified script. To retrieve
all fleets, leave this parameter empty.}

\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken}
to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this action. To
start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves a collection of fleet records for this AWS account. You can
filter the result set to find only those fleets that are deployed with a
specific build or script. Use the pagination parameters to retrieve
results in sequential pages.
}
\details{
Fleet records are not listed in a particular order.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Set Up Fleets}.

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\itemize{
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
}
\item Update fleets:
\itemize{
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
}
\item Manage fleet actions:
\itemize{
\item StartFleetActions
\item StopFleetActions
}
}
}
\section{Request syntax}{
\preformatted{svc$list_fleets(
  BuildId = "string",
  ScriptId = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
