% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_tags_for_resource}
\alias{gamelift_list_tags_for_resource}
\title{Retrieves all tags that are assigned to a GameLift resource}
\usage{
gamelift_list_tags_for_resource(ResourceARN)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name
(\href{https://docs.aws.amazon.com/AmazonS3/latest/userguide/s3-arn-format.html}{ARN})
that is assigned to and uniquely identifies the GameLift resource that
you want to retrieve tags for. GameLift resource ARNs are included in
the data object for the resource, which can be retrieved by calling a
List or Describe operation for the resource type.}
}
\value{
A list with the following syntax:\preformatted{list(
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}
\description{
Retrieves all tags that are assigned to a GameLift resource. Resource
tags are used to organize AWS resources for a range of purposes. This
operation handles the permissions necessary to manage tags for the
following GameLift resource types:
\itemize{
\item Build
\item Script
\item Fleet
\item Alias
\item GameSessionQueue
\item MatchmakingConfiguration
\item MatchmakingRuleSet
}

To list tags for a resource, specify the unique ARN value for the
resource.

\strong{Learn more}

\href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
in the \emph{AWS General Reference}

\href{https://d1.awsstatic.com/whitepapers/aws-tagging-best-practices.pdf}{AWS Tagging Strategies}

\strong{Related actions}

\code{\link[=gamelift_tag_resource]{tag_resource}} |
\code{\link[=gamelift_untag_resource]{untag_resource}} |
\code{\link[=gamelift_list_tags_for_resource]{list_tags_for_resource}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$list_tags_for_resource(
  ResourceARN = "string"
)
}
}

\keyword{internal}
