% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_register_game_server}
\alias{gamelift_register_game_server}
\title{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups}
\usage{
gamelift_register_game_server(GameServerGroupName, GameServerId,
  InstanceId, ConnectionInfo, GameServerData)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group where the game server is
running. Use either the GameServerGroup name or ARN value.}

\item{GameServerId}{[required] A custom string that uniquely identifies the game server to register.
Game server IDs are developer-defined and must be unique across all game
server groups in your AWS account.}

\item{InstanceId}{[required] The unique identifier for the instance where the game server is running.
This ID is available in the instance metadata. EC2 instance IDs use a
17-character format, for example: \verb{i-1234567890abcdef0}.}

\item{ConnectionInfo}{Information that is needed to make inbound client connections to the
game server. This might include the IP address and port, DNS name, and
other information.}

\item{GameServerData}{A set of custom game server properties, formatted as a single string
value. This data is passed to a game client or service when it requests
information on game servers using
\code{\link[=gamelift_list_game_servers]{list_game_servers}} or
\code{\link[=gamelift_claim_game_server]{claim_game_server}}.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GameServer = list(
    GameServerGroupName = "string",
    GameServerGroupArn = "string",
    GameServerId = "string",
    InstanceId = "string",
    ConnectionInfo = "string",
    GameServerData = "string",
    ClaimStatus = "CLAIMED",
    UtilizationStatus = "AVAILABLE"|"UTILIZED",
    RegistrationTime = as.POSIXct(
      "2015-01-01"
    ),
    LastClaimTime = as.POSIXct(
      "2015-01-01"
    ),
    LastHealthCheckTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
\strong{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups.}

Creates a new game server resource and notifies GameLift FleetIQ that
the game server is ready to host gameplay and players. This operation is
called by a game server process that is running on an instance in a game
server group. Registering game servers enables GameLift FleetIQ to track
available game servers and enables game clients and services to claim a
game server for a new game session.

To register a game server, identify the game server group and instance
where the game server is running, and provide a unique identifier for
the game server. You can also include connection and game server data.
When a game client or service requests a game server by calling
\code{\link[=gamelift_claim_game_server]{claim_game_server}}, this information is
returned in the response.

Once a game server is successfully registered, it is put in status
\code{AVAILABLE}. A request to register a game server may fail if the
instance it is running on is in the process of shutting down as part of
instance balancing or scale-down activity.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_register_game_server]{register_game_server}}
\item \code{\link[=gamelift_list_game_servers]{list_game_servers}}
\item \code{\link[=gamelift_claim_game_server]{claim_game_server}}
\item \code{\link[=gamelift_describe_game_server]{describe_game_server}}
\item \code{\link[=gamelift_update_game_server]{update_game_server}}
\item \code{\link[=gamelift_deregister_game_server]{deregister_game_server}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$register_game_server(
  GameServerGroupName = "string",
  GameServerId = "string",
  InstanceId = "string",
  ConnectionInfo = "string",
  GameServerData = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
