% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_request_upload_credentials}
\alias{gamelift_request_upload_credentials}
\title{Retrieves a fresh set of credentials for use when uploading a new set of
game build files to Amazon GameLift's Amazon S3}
\usage{
gamelift_request_upload_credentials(BuildId)
}
\arguments{
\item{BuildId}{[required] A unique identifier for a build to get credentials for. You can use
either the build ID or ARN value.}
}
\description{
Retrieves a fresh set of credentials for use when uploading a new set of
game build files to Amazon GameLift\'s Amazon S3. This is done as part
of the build creation process; see CreateBuild.
}
\details{
To request new credentials, specify the build ID as returned with an
initial \code{CreateBuild} request. If successful, a new set of credentials
are returned, along with the S3 storage location associated with the
build ID.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-build-cli-uploading.html#gamelift-build-cli-uploading-create-build}{Create a Build with Files in S3}

\strong{Related operations}
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\section{Request syntax}{
\preformatted{svc$request_upload_credentials(
  BuildId = "string"
)
}
}

\keyword{internal}
