% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_resolve_alias}
\alias{gamelift_resolve_alias}
\title{Retrieves the fleet ID that an alias is currently pointing to}
\usage{
gamelift_resolve_alias(AliasId)
}
\arguments{
\item{AliasId}{[required] The unique identifier of the alias that you want to retrieve a fleet ID
for. You can use either the alias ID or ARN value.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetId = "string",
  FleetArn = "string"
)
}
}
\description{
Retrieves the fleet ID that an alias is currently pointing to.
\itemize{
\item \code{\link[=gamelift_create_alias]{create_alias}}
\item \code{\link[=gamelift_list_aliases]{list_aliases}}
\item \code{\link[=gamelift_describe_alias]{describe_alias}}
\item \code{\link[=gamelift_update_alias]{update_alias}}
\item \code{\link[=gamelift_delete_alias]{delete_alias}}
\item \code{\link[=gamelift_resolve_alias]{resolve_alias}}
}
}
\section{Request syntax}{
\preformatted{svc$resolve_alias(
  AliasId = "string"
)
}
}

\keyword{internal}
