% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_start_game_session_placement}
\alias{gamelift_start_game_session_placement}
\title{Places a request for a new game session in a queue (see
CreateGameSessionQueue)}
\usage{
gamelift_start_game_session_placement(PlacementId, GameSessionQueueName,
  GameProperties, MaximumPlayerSessionCount, GameSessionName,
  PlayerLatencies, DesiredPlayerSessions, GameSessionData)
}
\arguments{
\item{PlacementId}{[required] A unique identifier to assign to the new game session placement. This
value is developer-defined. The value must be unique across all Regions
and cannot be reused unless you are resubmitting a canceled or timed-out
placement request.}

\item{GameSessionQueueName}{[required] Name of the queue to use to place the new game session. You can use
either the queue name or ARN value.}

\item{GameProperties}{Set of custom properties for a game session, formatted as key:value
pairs. These properties are passed to a game server process in the
GameSession object with a request to start a new game session (see
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).}

\item{MaximumPlayerSessionCount}{[required] The maximum number of players that can be connected simultaneously to
the game session.}

\item{GameSessionName}{A descriptive label that is associated with a game session. Session
names do not need to be unique.}

\item{PlayerLatencies}{Set of values, expressed in milliseconds, indicating the amount of
latency that a player experiences when connected to AWS Regions. This
information is used to try to place the new game session where it can
offer the best possible gameplay experience for the players.}

\item{DesiredPlayerSessions}{Set of information on each player to create a player session for.}

\item{GameSessionData}{Set of custom game session properties, formatted as a single string
value. This data is passed to a game server process in the GameSession
object with a request to start a new game session (see \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).}
}
\description{
Places a request for a new game session in a queue (see
\code{\link[=gamelift_create_game_session_queue]{create_game_session_queue}}). When
processing a placement request, Amazon GameLift searches for available
resources on the queue's destinations, scanning each until it finds
resources or the placement request times out.

A game session placement request can also request player sessions. When
a new game session is successfully created, Amazon GameLift creates a
player session for each player included in the request.

When placing a game session, by default Amazon GameLift tries each fleet
in the order they are listed in the queue configuration. Ideally, a
queue's destinations are listed in preference order.

Alternatively, when requesting a game session with players, you can also
provide latency data for each player in relevant Regions. Latency data
indicates the performance lag a player experiences when connected to a
fleet in the Region. Amazon GameLift uses latency data to reorder the
list of destinations to place the game session in a Region with minimal
lag. If latency data is provided for multiple players, Amazon GameLift
calculates each Region's average lag for all players and reorders to get
the best game play across all players.

To place a new game session request, specify the following:
\itemize{
\item The queue name and a set of game session properties and settings
\item A unique ID (such as a UUID) for the placement. You use this ID to
track the status of the placement request
\item (Optional) A set of player data and a unique player ID for each
player that you are joining to the new game session (player data is
optional, but if you include it, you must also provide a unique ID
for each player)
\item Latency data for all players (if you want to optimize game play for
the players)
}

If successful, a new game session placement is created.

To track the status of a placement request, call
\code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
and check the request's status. If the status is \code{FULFILLED}, a new game
session has been created and a game session ARN and Region are
referenced. If the placement request times out, you can resubmit the
request or retry it with a different queue.
\itemize{
\item \code{\link[=gamelift_create_game_session]{create_game_session}}
\item \code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}}
\item \code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
\item \code{\link[=gamelift_search_game_sessions]{search_game_sessions}}
\item \code{\link[=gamelift_update_game_session]{update_game_session}}
\item \code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}}
\item Game session placements
\itemize{
\item \code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
\item \code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
\item \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
}
}
}
\section{Request syntax}{
\preformatted{svc$start_game_session_placement(
  PlacementId = "string",
  GameSessionQueueName = "string",
  GameProperties = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  MaximumPlayerSessionCount = 123,
  GameSessionName = "string",
  PlayerLatencies = list(
    list(
      PlayerId = "string",
      RegionIdentifier = "string",
      LatencyInMilliseconds = 123.0
    )
  ),
  DesiredPlayerSessions = list(
    list(
      PlayerId = "string",
      PlayerData = "string"
    )
  ),
  GameSessionData = "string"
)
}
}

\keyword{internal}
