% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_stop_matchmaking}
\alias{gamelift_stop_matchmaking}
\title{Cancels a matchmaking ticket or match backfill ticket that is currently
being processed}
\usage{
gamelift_stop_matchmaking(TicketId)
}
\arguments{
\item{TicketId}{[required] A unique identifier for a matchmaking ticket.}
}
\value{
An empty list.
}
\description{
Cancels a matchmaking ticket or match backfill ticket that is currently
being processed. To stop the matchmaking operation, specify the ticket
ID. If successful, work on the ticket is stopped, and the ticket status
is changed to \code{CANCELLED}.

This call is also used to turn off automatic backfill for an individual
game session. This is for game sessions that are created with a
matchmaking configuration that has automatic backfill enabled. The
ticket ID is included in the \code{MatchmakerData} of an updated game session
object, which is provided to the game server.

If the operation is successful, the service sends back an empty JSON
struct with the HTTP 200 response (not an empty HTTP body).

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-client.html}{Add FlexMatch to a game client}

\strong{Related actions}

\code{\link[=gamelift_start_matchmaking]{start_matchmaking}} |
\code{\link[=gamelift_describe_matchmaking]{describe_matchmaking}} |
\code{\link[=gamelift_stop_matchmaking]{stop_matchmaking}} |
\code{\link[=gamelift_accept_match]{accept_match}} |
\code{\link[=gamelift_start_match_backfill]{start_match_backfill}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$stop_matchmaking(
  TicketId = "string"
)
}
}

\keyword{internal}
