% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_tag_resource}
\alias{gamelift_tag_resource}
\title{Assigns a tag to a GameLift resource}
\usage{
gamelift_tag_resource(ResourceARN, Tags)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name
(\href{https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-arn-format.html}{ARN})
that is assigned to and uniquely identifies the GameLift resource that
you want to assign tags to. GameLift resource ARNs are included in the
data object for the resource, which can be retrieved by calling a List
or Describe operation for the resource type.}

\item{Tags}{[required] A list of one or more tags to assign to the specified GameLift resource.
Tags are developer-defined and structured as key-value pairs. The
maximum tag limit may be lower than stated. See \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
for actual tagging limits.}
}
\description{
Assigns a tag to a GameLift resource. AWS resource tags provide an
additional management tool set. You can use tags to organize resources,
create IAM permissions policies to manage access to groups of resources,
customize AWS cost breakdowns, etc. This operation handles the
permissions necessary to manage tags for the following GameLift resource
types:
}
\details{
\itemize{
\item Build
\item Script
\item Fleet
\item Alias
\item GameSessionQueue
\item MatchmakingConfiguration
\item MatchmakingRuleSet
}

To add a tag to a resource, specify the unique ARN value for the
resource and provide a tag list containing one or more tags. The
operation succeeds even if the list includes tags that are already
assigned to the specified resource.

\strong{Learn more}

\href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
in the \emph{AWS General Reference}

\href{https://d1.awsstatic.com/whitepapers/aws-tagging-best-practices.pdf}{AWS Tagging Strategies}

\strong{Related operations}
\itemize{
\item TagResource
\item UntagResource
\item ListTagsForResource
}
}
\section{Request syntax}{
\preformatted{svc$tag_resource(
  ResourceARN = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
