% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_untag_resource}
\alias{gamelift_untag_resource}
\title{Removes a tag that is assigned to a GameLift resource}
\usage{
gamelift_untag_resource(ResourceARN, TagKeys)
}
\arguments{
\item{ResourceARN}{[required] The Amazon Resource Name
(\href{https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-arn-format.html}{ARN})
that is assigned to and uniquely identifies the GameLift resource that
you want to remove tags from. GameLift resource ARNs are included in the
data object for the resource, which can be retrieved by calling a List
or Describe action for the resource type.}

\item{TagKeys}{[required] A list of one or more tags to remove from the specified GameLift
resource. Tags are developer-defined and structured as key-value pairs.}
}
\description{
Removes a tag that is assigned to a GameLift resource. Resource tags are
used to organize AWS resources for a range of purposes. This action
handles the permissions necessary to manage tags for the following
GameLift resource types:
}
\details{
\itemize{
\item Build
\item Script
\item Fleet
\item Alias
\item GameSessionQueue
\item MatchmakingConfiguration
\item MatchmakingRuleSet
}

To remove a tag from a resource, specify the unique ARN value for the
resource and provide a string list containing one or more tags to be
removed. This action succeeds even if the list includes tags that are
not currently assigned to the specified resource.

\strong{Learn more}

\href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
in the \emph{AWS General Reference}

\href{http://aws.amazon.com/answers/account-management/aws-tagging-strategies/}{AWS Tagging Strategies}

\strong{Related operations}
\itemize{
\item TagResource
\item UntagResource
\item ListTagsForResource
}
}
\section{Request syntax}{
\preformatted{svc$untag_resource(
  ResourceARN = "string",
  TagKeys = list(
    "string"
  )
)
}
}

\keyword{internal}
