% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_build}
\alias{gamelift_update_build}
\title{Updates metadata in a build resource, including the build name and
version}
\usage{
gamelift_update_build(BuildId, Name, Version)
}
\arguments{
\item{BuildId}{[required] A unique identifier for a build to update. You can use either the build
ID or ARN value.}

\item{Name}{A descriptive label that is associated with a build. Build names do not
need to be unique.}

\item{Version}{Version information that is associated with a build or script. Version
strings do not need to be unique.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Build = list(
    BuildId = "string",
    BuildArn = "string",
    Name = "string",
    Version = "string",
    Status = "INITIALIZED"|"READY"|"FAILED",
    SizeOnDisk = 123,
    OperatingSystem = "WINDOWS_2012"|"AMAZON_LINUX"|"AMAZON_LINUX_2",
    CreationTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Updates metadata in a build resource, including the build name and
version. To update the metadata, specify the build ID to update and
provide the new values. If successful, a build object containing the
updated metadata is returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-build-intro.html}{Upload a Custom Server Build}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_build]{create_build}}
\item \code{\link[=gamelift_list_builds]{list_builds}}
\item \code{\link[=gamelift_describe_build]{describe_build}}
\item \code{\link[=gamelift_update_build]{update_build}}
\item \code{\link[=gamelift_delete_build]{delete_build}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$update_build(
  BuildId = "string",
  Name = "string",
  Version = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
