% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_fleet_attributes}
\alias{gamelift_update_fleet_attributes}
\title{Updates a fleet's mutable attributes, including game session protection
and resource creation limits}
\usage{
gamelift_update_fleet_attributes(FleetId, Name, Description,
  NewGameSessionProtectionPolicy, ResourceCreationLimitPolicy,
  MetricGroups)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to update attribute metadata for. You
can use either the fleet ID or ARN value.}

\item{Name}{A descriptive label that is associated with a fleet. Fleet names do not
need to be unique.}

\item{Description}{A human-readable description of a fleet.}

\item{NewGameSessionProtectionPolicy}{The game session protection policy to apply to all new instances created
in this fleet. Instances that already exist are not affected. You can
set protection for individual instances using
\code{\link[=gamelift_update_game_session]{update_game_session}}.
\itemize{
\item \strong{NoProtection} -- The game session can be terminated during a
scale-down event.
\item \strong{FullProtection} -- If the game session is in an \code{ACTIVE} status,
it cannot be terminated during a scale-down event.
}}

\item{ResourceCreationLimitPolicy}{Policy settings that limit the number of game sessions an individual
player can create over a span of time.}

\item{MetricGroups}{The name of a metric group to add this fleet to. Use a metric group in
Amazon CloudWatch to aggregate the metrics from multiple fleets. Provide
an existing metric group name, or create a new metric group by providing
a new name. A fleet can only be in one metric group at a time.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetId = "string"
)
}
}
\description{
Updates a fleet's mutable attributes, including game session protection
and resource creation limits.

To update fleet attributes, specify the fleet ID and the property values
that you want to change.

If successful, an updated \code{FleetAttributes} object is returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift fleets}

\strong{Related actions}

\code{\link[=gamelift_create_fleet_locations]{create_fleet_locations}} |
\code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}} |
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}} |
\code{\link[=gamelift_update_fleet_port_settings]{update_fleet_port_settings}} |
\code{\link[=gamelift_update_runtime_configuration]{update_runtime_configuration}}
| \code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}} |
\code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} |
\code{\link[=gamelift_put_scaling_policy]{put_scaling_policy}} |
\code{\link[=gamelift_delete_fleet]{delete_fleet}} |
\code{\link[=gamelift_delete_fleet_locations]{delete_fleet_locations}} |
\code{\link[=gamelift_delete_scaling_policy]{delete_scaling_policy}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$update_fleet_attributes(
  FleetId = "string",
  Name = "string",
  Description = "string",
  NewGameSessionProtectionPolicy = "NoProtection"|"FullProtection",
  ResourceCreationLimitPolicy = list(
    NewGameSessionsPerCreator = 123,
    PolicyPeriodInMinutes = 123
  ),
  MetricGroups = list(
    "string"
  )
)
}
}

\keyword{internal}
