% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_game_server_group}
\alias{gamelift_update_game_server_group}
\title{THIS ACTION IS PART OF AMAZON GAMELIFT FLEETIQ WITH GAME SERVER GROUPS,
WHICH IS IN PREVIEW RELEASE AND IS SUBJECT TO CHANGE}
\usage{
gamelift_update_game_server_group(GameServerGroupName, RoleArn,
  InstanceDefinitions, GameServerProtectionPolicy, BalancingStrategy)
}
\arguments{
\item{GameServerGroupName}{[required] The unique identifier of the game server group to update. Use either the
GameServerGroup name or ARN value.}

\item{RoleArn}{The Amazon Resource Name
(\href{https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-arn-format.html}{ARN})
for an IAM role that allows Amazon GameLift to access your EC2 Auto
Scaling groups. The submitted role is validated to ensure that it
contains the necessary permissions for game server groups.}

\item{InstanceDefinitions}{An updated list of EC2 instance types to use when creating instances in
the group. The instance definition must specify instance types that are
supported by GameLift FleetIQ, and must include at least two instance
types. This updated list replaces the entire current list of instance
definitions for the game server group. For more information on instance
types, see \href{https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/instance-types.html}{EC2 Instance Types}
in the \emph{Amazon EC2 User Guide}..}

\item{GameServerProtectionPolicy}{A flag that indicates whether instances in the game server group are
protected from early termination. Unprotected instances that have active
game servers running may by terminated during a scale-down event,
causing players to be dropped from the game. Protected instances cannot
be terminated while there are active game servers running. An exception
to this is Spot Instances, which may be terminated by AWS regardless of
protection status. This property is set to NO\\_PROTECTION by default.}

\item{BalancingStrategy}{The fallback balancing method to use for the game server group when Spot
instances in a Region become unavailable or are not viable for game
hosting. Once triggered, this method remains active until Spot instances
can once again be used. Method options include:
\itemize{
\item SPOT\\_ONLY \\-- If Spot instances are unavailable, the game server
group provides no hosting capacity. No new instances are started,
and the existing nonviable Spot instances are terminated (once
current gameplay ends) and not replaced.
\item SPOT\\_PREFERRED \\-- If Spot instances are unavailable, the game
server group continues to provide hosting capacity by using
On-Demand instances. Existing nonviable Spot instances are
terminated (once current gameplay ends) and replaced with new
On-Demand instances.
}}
}
\description{
\strong{This action is part of Amazon GameLift FleetIQ with game server
groups, which is in preview release and is subject to change.}
}
\details{
Updates GameLift FleetIQ-specific properties for a game server group.
These properties include instance rebalancing and game server
protection. Many Auto Scaling group properties are updated directly.
These include autoscaling policies, minimum/maximum/desired instance
counts, and launch template.

To update the game server group, specify the game server group ID and
provide the updated values.

Updated properties are validated to ensure that GameLift FleetIQ can
continue to perform its core instance rebalancing activity. When you
change Auto Scaling group properties directly and the changes cause
errors with GameLift FleetIQ activities, an alert is sent.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-intro.html}{GameLift FleetIQ Guide}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-asgroups.html}{Updating a GameLift FleetIQ-Linked Auto Scaling Group}

\strong{Related operations}
\itemize{
\item CreateGameServerGroup
\item ListGameServerGroups
\item DescribeGameServerGroup
\item UpdateGameServerGroup
\item DeleteGameServerGroup
\item ResumeGameServerGroup
\item SuspendGameServerGroup
}
}
\section{Request syntax}{
\preformatted{svc$update_game_server_group(
  GameServerGroupName = "string",
  RoleArn = "string",
  InstanceDefinitions = list(
    list(
      InstanceType = "c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge",
      WeightedCapacity = "string"
    )
  ),
  GameServerProtectionPolicy = "NO_PROTECTION"|"FULL_PROTECTION",
  BalancingStrategy = "SPOT_ONLY"|"SPOT_PREFERRED"
)
}
}

\keyword{internal}
