% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_game_session}
\alias{gamelift_update_game_session}
\title{Updates the mutable properties of a game session}
\usage{
gamelift_update_game_session(GameSessionId, MaximumPlayerSessionCount,
  Name, PlayerSessionCreationPolicy, ProtectionPolicy)
}
\arguments{
\item{GameSessionId}{[required] A unique identifier for the game session to update.}

\item{MaximumPlayerSessionCount}{The maximum number of players that can be connected simultaneously to
the game session.}

\item{Name}{A descriptive label that is associated with a game session. Session
names do not need to be unique.}

\item{PlayerSessionCreationPolicy}{A policy that determines whether the game session is accepting new
players.}

\item{ProtectionPolicy}{Game session protection policy to apply to this game session only.
\itemize{
\item \strong{NoProtection} -- The game session can be terminated during a
scale-down event.
\item \strong{FullProtection} -- If the game session is in an \code{ACTIVE} status,
it cannot be terminated during a scale-down event.
}}
}
\value{
A list with the following syntax:\preformatted{list(
  GameSession = list(
    GameSessionId = "string",
    Name = "string",
    FleetId = "string",
    FleetArn = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    TerminationTime = as.POSIXct(
      "2015-01-01"
    ),
    CurrentPlayerSessionCount = 123,
    MaximumPlayerSessionCount = 123,
    Status = "ACTIVE"|"ACTIVATING"|"TERMINATED"|"TERMINATING"|"ERROR",
    StatusReason = "INTERRUPTED",
    GameProperties = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    IpAddress = "string",
    DnsName = "string",
    Port = 123,
    PlayerSessionCreationPolicy = "ACCEPT_ALL"|"DENY_ALL",
    CreatorId = "string",
    GameSessionData = "string",
    MatchmakerData = "string",
    Location = "string"
  )
)
}
}
\description{
Updates the mutable properties of a game session.

To update a game session, specify the game session ID and the values you
want to change.

If successful, the updated \code{GameSession} object is returned.

\strong{Related actions}

\code{\link[=gamelift_create_game_session]{create_game_session}} |
\code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}} |
\code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
| \code{\link[=gamelift_search_game_sessions]{search_game_sessions}} |
\code{\link[=gamelift_update_game_session]{update_game_session}} |
\code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}} |
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
|
\code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
| \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$update_game_session(
  GameSessionId = "string",
  MaximumPlayerSessionCount = 123,
  Name = "string",
  PlayerSessionCreationPolicy = "ACCEPT_ALL"|"DENY_ALL",
  ProtectionPolicy = "NoProtection"|"FullProtection"
)
}
}

\keyword{internal}
