% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_matchmaking_configuration}
\alias{gamelift_update_matchmaking_configuration}
\title{Updates settings for a FlexMatch matchmaking configuration}
\usage{
gamelift_update_matchmaking_configuration(Name, Description,
  GameSessionQueueArns, RequestTimeoutSeconds, AcceptanceTimeoutSeconds,
  AcceptanceRequired, RuleSetName, NotificationTarget,
  AdditionalPlayerCount, CustomEventData, GameProperties, GameSessionData,
  BackfillMode)
}
\arguments{
\item{Name}{[required] Unique identifier for a matchmaking configuration to update.}

\item{Description}{Descriptive label that is associated with matchmaking configuration.}

\item{GameSessionQueueArns}{Amazon Resource Name
(\href{https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-arn-format.html}{ARN})
that is assigned to a game session queue and uniquely identifies it.
Format is
\code{arn:aws:gamelift:&lt;region&gt;:&lt;aws account&gt;:gamesessionqueue/&lt;queue name&gt;}.
These queues are used when placing game sessions for matches that are
created with this matchmaking configuration. Queues can be located in
any region.}

\item{RequestTimeoutSeconds}{Maximum duration, in seconds, that a matchmaking ticket can remain in
process before timing out. Requests that fail due to timing out can be
resubmitted as needed.}

\item{AcceptanceTimeoutSeconds}{Length of time (in seconds) to wait for players to accept a proposed
match. If any player rejects the match or fails to accept before the
timeout, the ticket continues to look for an acceptable match.}

\item{AcceptanceRequired}{Flag that determines whether a match that was created with this
configuration must be accepted by the matched players. To require
acceptance, set to TRUE.}

\item{RuleSetName}{Unique identifier for a matchmaking rule set to use with this
configuration. A matchmaking configuration can only use rule sets that
are defined in the same region.}

\item{NotificationTarget}{SNS topic ARN that is set up to receive matchmaking notifications. See
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/match-notification.html}{Setting up Notifications for Matchmaking}
for more information.}

\item{AdditionalPlayerCount}{Number of player slots in a match to keep open for future players. For
example, if the configuration's rule set specifies a match for a single
12-person team, and the additional player count is set to 2, only 10
players are selected for the match.}

\item{CustomEventData}{Information to add to all events related to the matchmaking
configuration.}

\item{GameProperties}{Set of custom properties for a game session, formatted as key:value
pairs. These properties are passed to a game server process in the
GameSession object with a request to start a new game session (see
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).
This information is added to the new GameSession object that is created
for a successful match.}

\item{GameSessionData}{Set of custom game session properties, formatted as a single string
value. This data is passed to a game server process in the GameSession
object with a request to start a new game session (see \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).
This information is added to the new GameSession object that is created
for a successful match.}

\item{BackfillMode}{Method used to backfill game sessions created with this matchmaking
configuration. Specify MANUAL when your game manages backfill requests
manually or does not use the match backfill feature. Specify AUTOMATIC
to have GameLift create a StartMatchBackfill request whenever a game
session has one or more open slots. Learn more about manual and
automatic backfill in \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/match-backfill.html}{Backfill Existing Games with FlexMatch}.}
}
\description{
Updates settings for a FlexMatch matchmaking configuration. These
changes affect all matches and game sessions that are created after the
update. To update settings, specify the configuration name to be updated
and provide the new settings.
}
\details{
\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/match-configuration.html}{Design a FlexMatch Matchmaker}

\strong{Related operations}
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Request syntax}{
\preformatted{svc$update_matchmaking_configuration(
  Name = "string",
  Description = "string",
  GameSessionQueueArns = list(
    "string"
  ),
  RequestTimeoutSeconds = 123,
  AcceptanceTimeoutSeconds = 123,
  AcceptanceRequired = TRUE|FALSE,
  RuleSetName = "string",
  NotificationTarget = "string",
  AdditionalPlayerCount = 123,
  CustomEventData = "string",
  GameProperties = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  GameSessionData = "string",
  BackfillMode = "AUTOMATIC"|"MANUAL"
)
}
}

\keyword{internal}
