% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_runtime_configuration}
\alias{gamelift_update_runtime_configuration}
\title{Updates the current runtime configuration for the specified fleet, which
tells GameLift how to launch server processes on all instances in the
fleet}
\usage{
gamelift_update_runtime_configuration(FleetId, RuntimeConfiguration)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to update runtime configuration for.
You can use either the fleet ID or ARN value.}

\item{RuntimeConfiguration}{[required] Instructions for launching server processes on each instance in the
fleet. Server processes run either a custom game build executable or a
Realtime Servers script. The runtime configuration lists the types of
server processes to run on an instance, how to launch them, and the
number of processes to run concurrently.}
}
\value{
A list with the following syntax:\preformatted{list(
  RuntimeConfiguration = list(
    ServerProcesses = list(
      list(
        LaunchPath = "string",
        Parameters = "string",
        ConcurrentExecutions = 123
      )
    ),
    MaxConcurrentGameSessionActivations = 123,
    GameSessionActivationTimeoutSeconds = 123
  )
)
}
}
\description{
Updates the current runtime configuration for the specified fleet, which
tells GameLift how to launch server processes on all instances in the
fleet. You can update a fleet's runtime configuration at any time after
the fleet is created; it does not need to be in \code{ACTIVE} status.

To update runtime configuration, specify the fleet ID and provide a
\code{RuntimeConfiguration} with an updated set of server process
configurations.

If successful, the fleet's runtime configuration settings are updated.
Each instance in the fleet regularly checks for and retrieves updated
runtime configurations. Instances immediately begin complying with the
new configuration by launching new server processes or not replacing
existing processes when they shut down. Updating a fleet's runtime
configuration never affects existing server processes.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift fleets}

\strong{Related actions}

\code{\link[=gamelift_create_fleet_locations]{create_fleet_locations}} |
\code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}} |
\code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}} |
\code{\link[=gamelift_update_fleet_port_settings]{update_fleet_port_settings}} |
\code{\link[=gamelift_update_runtime_configuration]{update_runtime_configuration}}
| \code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}} |
\code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} |
\code{\link[=gamelift_put_scaling_policy]{put_scaling_policy}} |
\code{\link[=gamelift_delete_fleet]{delete_fleet}} |
\code{\link[=gamelift_delete_fleet_locations]{delete_fleet_locations}} |
\code{\link[=gamelift_delete_scaling_policy]{delete_scaling_policy}} | \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$update_runtime_configuration(
  FleetId = "string",
  RuntimeConfiguration = list(
    ServerProcesses = list(
      list(
        LaunchPath = "string",
        Parameters = "string",
        ConcurrentExecutions = 123
      )
    ),
    MaxConcurrentGameSessionActivations = 123,
    GameSessionActivationTimeoutSeconds = 123
  )
)
}
}

\keyword{internal}
