% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_validate_matchmaking_rule_set}
\alias{gamelift_validate_matchmaking_rule_set}
\title{Validates the syntax of a matchmaking rule or rule set}
\usage{
gamelift_validate_matchmaking_rule_set(RuleSetBody)
}
\arguments{
\item{RuleSetBody}{[required] A collection of matchmaking rules to validate, formatted as a JSON
string.}
}
\value{
A list with the following syntax:\preformatted{list(
  Valid = TRUE|FALSE
)
}
}
\description{
Validates the syntax of a matchmaking rule or rule set. This operation
checks that the rule set is using syntactically correct JSON and that it
conforms to allowed property expressions. To validate syntax, provide a
rule set JSON string.

\strong{Learn more}
\itemize{
\item \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-rulesets.html}{Build a rule set}
}

\strong{Related actions}

\code{\link[=gamelift_create_matchmaking_configuration]{create_matchmaking_configuration}}
|
\code{\link[=gamelift_describe_matchmaking_configurations]{describe_matchmaking_configurations}}
|
\code{\link[=gamelift_update_matchmaking_configuration]{update_matchmaking_configuration}}
|
\code{\link[=gamelift_delete_matchmaking_configuration]{delete_matchmaking_configuration}}
| \code{\link[=gamelift_create_matchmaking_rule_set]{create_matchmaking_rule_set}}
|
\code{\link[=gamelift_describe_matchmaking_rule_sets]{describe_matchmaking_rule_sets}}
|
\code{\link[=gamelift_validate_matchmaking_rule_set]{validate_matchmaking_rule_set}}
| \code{\link[=gamelift_delete_matchmaking_rule_set]{delete_matchmaking_rule_set}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$validate_matchmaking_rule_set(
  RuleSetBody = "string"
)
}
}

\keyword{internal}
