% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_accept_match}
\alias{gamelift_accept_match}
\title{Registers a player's acceptance or rejection of a proposed FlexMatch
match}
\usage{
gamelift_accept_match(TicketId, PlayerIds, AcceptanceType)
}
\arguments{
\item{TicketId}{[required] Unique identifier for a matchmaking ticket. The ticket must be in status
\code{REQUIRES_ACCEPTANCE}; otherwise this request will fail.}

\item{PlayerIds}{[required] Unique identifier for a player delivering the response. This parameter
can include one or multiple player IDs.}

\item{AcceptanceType}{[required] Player response to the proposed match.}
}
\description{
Registers a player's acceptance or rejection of a proposed FlexMatch
match. A matchmaking configuration may require player acceptance; if so,
then matches built with that configuration cannot be completed unless
all players accept the proposed match within a specified time limit.
}
\details{
When FlexMatch builds a match, all the matchmaking tickets involved in
the proposed match are placed into status \code{REQUIRES_ACCEPTANCE}. This is
a trigger for your game to get acceptance from all players in the
ticket. Acceptances are only valid for tickets when they are in this
status; all other acceptances result in an error.

To register acceptance, specify the ticket ID, a response, and one or
more players. Once all players have registered acceptance, the
matchmaking tickets advance to status \code{PLACING}, where a new game
session is created for the match.

If any player rejects the match, or if acceptances are not received
before a specified timeout, the proposed match is dropped. The
matchmaking tickets are then handled in one of two ways: For tickets
where all players accepted the match, the ticket status is returned to
\code{SEARCHING} to find a new match. For tickets where one or more players
failed to accept the match, the ticket status is set to \code{FAILED}, and
processing is terminated. A new matchmaking request for these players
can be submitted as needed.
\itemize{
\item StartMatchmaking
\item DescribeMatchmaking
\item StopMatchmaking
\item AcceptMatch
\item StartMatchBackfill
}
}
\section{Request syntax}{
\preformatted{svc$accept_match(
  TicketId = "string",
  PlayerIds = list(
    "string"
  ),
  AcceptanceType = "ACCEPT"|"REJECT"
)
}
}

\keyword{internal}
