% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_claim_game_server}
\alias{gamelift_claim_game_server}
\title{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups}
\usage{
gamelift_claim_game_server(GameServerGroupName, GameServerId,
  GameServerData)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group where the game server is
running. Use either the GameServerGroup name or ARN value.. If you are
not specifying a game server to claim, this value identifies where you
want GameLift FleetIQ to look for an available game server to claim.}

\item{GameServerId}{A custom string that uniquely identifies the game server to claim. If
this parameter is left empty, GameLift FleetIQ searches for an available
game server in the specified game server group.}

\item{GameServerData}{A set of custom game server properties, formatted as a single string
value. This data is passed to a game client or service when it requests
information on game servers using
\code{\link[=gamelift_list_game_servers]{list_game_servers}} or
\code{\link[=gamelift_claim_game_server]{claim_game_server}}.}
}
\value{
A list with the following syntax:\preformatted{list(
  GameServer = list(
    GameServerGroupName = "string",
    GameServerGroupArn = "string",
    GameServerId = "string",
    InstanceId = "string",
    ConnectionInfo = "string",
    GameServerData = "string",
    ClaimStatus = "CLAIMED",
    UtilizationStatus = "AVAILABLE"|"UTILIZED",
    RegistrationTime = as.POSIXct(
      "2015-01-01"
    ),
    LastClaimTime = as.POSIXct(
      "2015-01-01"
    ),
    LastHealthCheckTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
\strong{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups.}

Locates an available game server and temporarily reserves it to host
gameplay and players. This operation is called from a game client or
client service (such as a matchmaker) to request hosting resources for a
new game session. In response, GameLift FleetIQ locates an available
game server, places it in \code{CLAIMED} status for 60 seconds, and returns
connection information that players can use to connect to the game
server.

To claim a game server, identify a game server group. You can also
specify a game server ID, although this approach bypasses GameLift
FleetIQ placement optimization. Optionally, include game data to pass to
the game server at the start of a game session, such as a game map or
player information.

When a game server is successfully claimed, connection information is
returned. A claimed game server's utilization status remains \code{AVAILABLE}
while the claim status is set to \code{CLAIMED} for up to 60 seconds. This
time period gives the game server time to update its status to
\code{UTILIZED} (using \code{\link[=gamelift_update_game_server]{update_game_server}})
once players join. If the game server's status is not updated within 60
seconds, the game server reverts to unclaimed status and is available to
be claimed by another request. The claim time period is a fixed value
and is not configurable.

If you try to claim a specific game server, this request will fail in
the following cases:
\itemize{
\item If the game server utilization status is \code{UTILIZED}.
\item If the game server claim status is \code{CLAIMED}.
}

When claiming a specific game server, this request will succeed even if
the game server is running on an instance in \code{DRAINING} status. To avoid
this, first check the instance status by calling
\code{\link[=gamelift_describe_game_server_instances]{describe_game_server_instances}}.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_register_game_server]{register_game_server}}
\item \code{\link[=gamelift_list_game_servers]{list_game_servers}}
\item \code{\link[=gamelift_claim_game_server]{claim_game_server}}
\item \code{\link[=gamelift_describe_game_server]{describe_game_server}}
\item \code{\link[=gamelift_update_game_server]{update_game_server}}
\item \code{\link[=gamelift_deregister_game_server]{deregister_game_server}}
}
}
\section{Request syntax}{
\preformatted{svc$claim_game_server(
  GameServerGroupName = "string",
  GameServerId = "string",
  GameServerData = "string"
)
}
}

\keyword{internal}
