% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_alias}
\alias{gamelift_create_alias}
\title{Creates an alias for a fleet}
\usage{
gamelift_create_alias(Name, Description, RoutingStrategy, Tags)
}
\arguments{
\item{Name}{[required] A descriptive label that is associated with an alias. Alias names do not
need to be unique.}

\item{Description}{A human-readable description of the alias.}

\item{RoutingStrategy}{[required] The routing configuration, including routing type and fleet target, for
the alias.}

\item{Tags}{A list of labels to assign to the new alias resource. Tags are
developer-defined key-value pairs. Tagging AWS resources are useful for
resource management, access management and cost allocation. For more
information, see \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
in the \emph{AWS General Reference}. Once the resource is created, you can
use \code{\link[=gamelift_tag_resource]{tag_resource}},
\code{\link[=gamelift_untag_resource]{untag_resource}}, and
\code{\link[=gamelift_list_tags_for_resource]{list_tags_for_resource}} to add,
remove, and view tags. The maximum tag limit may be lower than stated.
See the AWS General Reference for actual tagging limits.}
}
\description{
Creates an alias for a fleet. In most situations, you can use an alias
ID in place of a fleet ID. An alias provides a level of abstraction for
a fleet that is useful when redirecting player traffic from one fleet to
another, such as when updating your game build.

Amazon GameLift supports two types of routing strategies for aliases:
simple and terminal. A simple alias points to an active fleet. A
terminal alias is used to display messaging or link to a URL instead of
routing players to an active fleet. For example, you might use a
terminal alias when a game version is no longer supported and you want
to direct players to an upgrade site.

To create a fleet alias, specify an alias name, routing strategy, and
optional description. Each simple alias can point to only one fleet, but
a fleet can have multiple aliases. If successful, a new alias record is
returned, including an alias ID and an ARN. You can reassign an alias to
another fleet by calling \code{\link[=gamelift_update_alias]{update_alias}}.
\itemize{
\item \code{\link[=gamelift_create_alias]{create_alias}}
\item \code{\link[=gamelift_list_aliases]{list_aliases}}
\item \code{\link[=gamelift_describe_alias]{describe_alias}}
\item \code{\link[=gamelift_update_alias]{update_alias}}
\item \code{\link[=gamelift_delete_alias]{delete_alias}}
\item \code{\link[=gamelift_resolve_alias]{resolve_alias}}
}
}
\section{Request syntax}{
\preformatted{svc$create_alias(
  Name = "string",
  Description = "string",
  RoutingStrategy = list(
    Type = "SIMPLE"|"TERMINAL",
    FleetId = "string",
    Message = "string"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
