% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_fleet}
\alias{gamelift_create_fleet}
\title{Creates a new fleet to run your game servers}
\usage{
gamelift_create_fleet(Name, Description, BuildId, ScriptId,
  ServerLaunchPath, ServerLaunchParameters, LogPaths, EC2InstanceType,
  EC2InboundPermissions, NewGameSessionProtectionPolicy,
  RuntimeConfiguration, ResourceCreationLimitPolicy, MetricGroups,
  PeerVpcAwsAccountId, PeerVpcId, FleetType, InstanceRoleArn,
  CertificateConfiguration, Tags)
}
\arguments{
\item{Name}{[required] A descriptive label that is associated with a fleet. Fleet names do not
need to be unique.}

\item{Description}{A human-readable description of a fleet.}

\item{BuildId}{A unique identifier for a build to be deployed on the new fleet. You can
use either the build ID or ARN value. The custom game server build must
have been successfully uploaded to Amazon GameLift and be in a \code{READY}
status. This fleet setting cannot be changed once the fleet is created.}

\item{ScriptId}{A unique identifier for a Realtime script to be deployed on the new
fleet. You can use either the script ID or ARN value. The Realtime
script must have been successfully uploaded to Amazon GameLift. This
fleet setting cannot be changed once the fleet is created.}

\item{ServerLaunchPath}{This parameter is no longer used. Instead, specify a server launch path
using the \code{RuntimeConfiguration} parameter. Requests that specify a
server launch path and launch parameters instead of a runtime
configuration will continue to work.}

\item{ServerLaunchParameters}{This parameter is no longer used. Instead, specify server launch
parameters in the \code{RuntimeConfiguration} parameter. (Requests that
specify a server launch path and launch parameters instead of a runtime
configuration will continue to work.)}

\item{LogPaths}{This parameter is no longer used. Instead, to specify where Amazon
GameLift should store log files once a server process shuts down, use
the Amazon GameLift server API \code{ProcessReady()} and specify one or more
directory paths in \code{logParameters}. See more information in the \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/#gamelift-sdk-server-api-ref-dataypes-process}{Server API Reference}.}

\item{EC2InstanceType}{[required] The name of an EC2 instance type that is supported in Amazon GameLift. A
fleet instance type determines the computing resources of each instance
in the fleet, including CPU, memory, storage, and networking capacity.
Amazon GameLift supports the following EC2 instance types. See \href{https://aws.amazon.com/ec2/instance-types/}{Amazon EC2 Instance Types} for
detailed descriptions.}

\item{EC2InboundPermissions}{Range of IP addresses and port settings that permit inbound traffic to
access game sessions that are running on the fleet. For fleets using a
custom game build, this parameter is required before game sessions
running on the fleet can accept connections. For Realtime Servers
fleets, Amazon GameLift automatically sets TCP and UDP ranges for use by
the Realtime servers. You can specify multiple permission settings or
add more by updating the fleet.}

\item{NewGameSessionProtectionPolicy}{A game session protection policy to apply to all instances in this
fleet. If this parameter is not set, instances in this fleet default to
no protection. You can change a fleet's protection policy using
\code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}}, but this
change will only affect sessions created after the policy change. You
can also set protection for individual instances using
\code{\link[=gamelift_update_game_session]{update_game_session}}.
\itemize{
\item \strong{NoProtection} - The game session can be terminated during a
scale-down event.
\item \strong{FullProtection} - If the game session is in an \code{ACTIVE} status,
it cannot be terminated during a scale-down event.
}}

\item{RuntimeConfiguration}{Instructions for launching server processes on each instance in the
fleet. Server processes run either a custom game build executable or a
Realtime script. The runtime configuration defines the server
executables or launch script file, launch parameters, and the number of
processes to run concurrently on each instance. When creating a fleet,
the runtime configuration must have at least one server process
configuration; otherwise the request fails with an invalid request
exception. (This parameter replaces the parameters \code{ServerLaunchPath}
and \code{ServerLaunchParameters}, although requests that contain values for
these parameters instead of a runtime configuration will continue to
work.) This parameter is required unless the parameters
\code{ServerLaunchPath} and \code{ServerLaunchParameters} are defined. Runtime
configuration replaced these parameters, but fleets that use them will
continue to work.}

\item{ResourceCreationLimitPolicy}{A policy that limits the number of game sessions an individual player
can create over a span of time for this fleet.}

\item{MetricGroups}{The name of an Amazon CloudWatch metric group to add this fleet to. A
metric group aggregates the metrics for all fleets in the group. Specify
an existing metric group name, or provide a new name to create a new
metric group. A fleet can only be included in one metric group at a
time.}

\item{PeerVpcAwsAccountId}{A unique identifier for the AWS account with the VPC that you want to
peer your Amazon GameLift fleet with. You can find your account ID in
the AWS Management Console under account settings.}

\item{PeerVpcId}{A unique identifier for a VPC with resources to be accessed by your
Amazon GameLift fleet. The VPC must be in the same Region as your fleet.
To look up a VPC ID, use the VPC Dashboard in the AWS Management
Console. Learn more about VPC peering in \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/vpc-peering.html}{VPC Peering with Amazon GameLift Fleets}.}

\item{FleetType}{Indicates whether to use On-Demand instances or Spot instances for this
fleet. If empty, the default is \code{ON_DEMAND}. Both categories of
instances use identical hardware and configurations based on the
instance type selected for this fleet. Learn more about \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-ec2-instances.html#gamelift-ec2-instances-spot}{On-Demand versus Spot Instances}.}

\item{InstanceRoleArn}{A unique identifier for an AWS IAM role that manages access to your AWS
services. Fleets with an instance role ARN allow applications that are
running on the fleet's instances to assume the role. Learn more about
using on-box credentials for your game servers at \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-resources.html}{Access external resources from a game server}.
To call this operation with instance role ARN, you must have IAM
PassRole permissions. See \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-iam-policy-examples.html}{IAM policy examples for GameLift}.}

\item{CertificateConfiguration}{Indicates whether to generate a TLS/SSL certificate for the new fleet.
TLS certificates are used for encrypting traffic between game clients
and game servers running on GameLift. If this parameter is not
specified, the default value, DISABLED, is used. This fleet setting
cannot be changed once the fleet is created. Learn more at \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-howitworks.html#gamelift-howitworks-security}{Securing Client/Server Communication}.

Note: This feature requires the AWS Certificate Manager (ACM) service,
which is available in the AWS global partition but not in all other
partitions. When working in a partition that does not support this
feature, a request for a new fleet with certificate generation results
fails with a 4xx unsupported Region error.

Valid values include:
\itemize{
\item \strong{GENERATED} - Generate a TLS/SSL certificate for this fleet.
\item \strong{DISABLED} - (default) Do not generate a TLS/SSL certificate for
this fleet.
}}

\item{Tags}{A list of labels to assign to the new fleet resource. Tags are
developer-defined key-value pairs. Tagging AWS resources are useful for
resource management, access management and cost allocation. For more
information, see \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
in the \emph{AWS General Reference}. Once the resource is created, you can
use \code{\link[=gamelift_tag_resource]{tag_resource}},
\code{\link[=gamelift_untag_resource]{untag_resource}}, and
\code{\link[=gamelift_list_tags_for_resource]{list_tags_for_resource}} to add,
remove, and view tags. The maximum tag limit may be lower than stated.
See the AWS General Reference for actual tagging limits.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  FleetAttributes = list(
    FleetId = "string",
    FleetArn = "string",
    FleetType = "ON_DEMAND"|"SPOT",
    InstanceType = "t2.micro"|"t2.small"|"t2.medium"|"t2.large"|"c3.large"|"c3.xlarge"|"c3.2xlarge"|"c3.4xlarge"|"c3.8xlarge"|"c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"c5a.large"|"c5a.xlarge"|"c5a.2xlarge"|"c5a.4xlarge"|"c5a.8xlarge"|"c5a.12xlarge"|"c5a.16xlarge"|"c5a.24xlarge"|"r3.large"|"r3.xlarge"|"r3.2xlarge"|"r3.4xlarge"|"r3.8xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"r5a.large"|"r5a.xlarge"|"r5a.2xlarge"|"r5a.4xlarge"|"r5a.8xlarge"|"r5a.12xlarge"|"r5a.16xlarge"|"r5a.24xlarge"|"m3.medium"|"m3.large"|"m3.xlarge"|"m3.2xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge"|"m5a.large"|"m5a.xlarge"|"m5a.2xlarge"|"m5a.4xlarge"|"m5a.8xlarge"|"m5a.12xlarge"|"m5a.16xlarge"|"m5a.24xlarge",
    Description = "string",
    Name = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    TerminationTime = as.POSIXct(
      "2015-01-01"
    ),
    Status = "NEW"|"DOWNLOADING"|"VALIDATING"|"BUILDING"|"ACTIVATING"|"ACTIVE"|"DELETING"|"ERROR"|"TERMINATED",
    BuildId = "string",
    BuildArn = "string",
    ScriptId = "string",
    ScriptArn = "string",
    ServerLaunchPath = "string",
    ServerLaunchParameters = "string",
    LogPaths = list(
      "string"
    ),
    NewGameSessionProtectionPolicy = "NoProtection"|"FullProtection",
    OperatingSystem = "WINDOWS_2012"|"AMAZON_LINUX"|"AMAZON_LINUX_2",
    ResourceCreationLimitPolicy = list(
      NewGameSessionsPerCreator = 123,
      PolicyPeriodInMinutes = 123
    ),
    MetricGroups = list(
      "string"
    ),
    StoppedActions = list(
      "AUTO_SCALING"
    ),
    InstanceRoleArn = "string",
    CertificateConfiguration = list(
      CertificateType = "DISABLED"|"GENERATED"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a new fleet to run your game servers. whether they are custom
game builds or Realtime Servers with game-specific script. A fleet is a
set of Amazon Elastic Compute Cloud (Amazon EC2) instances, each of
which can host multiple game sessions. When creating a fleet, you choose
the hardware specifications, set some configuration options, and specify
the game server to deploy on the new fleet.

To create a new fleet, provide the following: (1) a fleet name, (2) an
EC2 instance type and fleet type (spot or on-demand), (3) the build ID
for your game build or script ID if using Realtime Servers, and (4) a
runtime configuration, which determines how game servers will run on
each instance in the fleet.

If the \code{\link[=gamelift_create_fleet]{create_fleet}} call is successful,
Amazon GameLift performs the following tasks. You can track the process
of a fleet by checking the fleet status or by monitoring fleet creation
events:
\itemize{
\item Creates a fleet resource. Status: \code{NEW}.
\item Begins writing events to the fleet event log, which can be accessed
in the Amazon GameLift console.
\item Sets the fleet's target capacity to 1 (desired instances), which
triggers Amazon GameLift to start one new EC2 instance.
\item Downloads the game build or Realtime script to the new instance and
installs it. Statuses: \code{DOWNLOADING}, \code{VALIDATING}, \code{BUILDING}.
\item Starts launching server processes on the instance. If the fleet is
configured to run multiple server processes per instance, Amazon
GameLift staggers each process launch by a few seconds. Status:
\code{ACTIVATING}.
\item Sets the fleet's status to \code{ACTIVE} as soon as one server process is
ready to host a game session.
}

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting Up Fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-creating-debug.html#fleets-creating-debug-creation}{Debug Fleet Creation Issues}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_fleet]{create_fleet}}
\item \code{\link[=gamelift_list_fleets]{list_fleets}}
\item \code{\link[=gamelift_delete_fleet]{delete_fleet}}
\item \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}}
\item \code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}}
\item \code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} or
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_fleet(
  Name = "string",
  Description = "string",
  BuildId = "string",
  ScriptId = "string",
  ServerLaunchPath = "string",
  ServerLaunchParameters = "string",
  LogPaths = list(
    "string"
  ),
  EC2InstanceType = "t2.micro"|"t2.small"|"t2.medium"|"t2.large"|"c3.large"|"c3.xlarge"|"c3.2xlarge"|"c3.4xlarge"|"c3.8xlarge"|"c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"c5a.large"|"c5a.xlarge"|"c5a.2xlarge"|"c5a.4xlarge"|"c5a.8xlarge"|"c5a.12xlarge"|"c5a.16xlarge"|"c5a.24xlarge"|"r3.large"|"r3.xlarge"|"r3.2xlarge"|"r3.4xlarge"|"r3.8xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"r5a.large"|"r5a.xlarge"|"r5a.2xlarge"|"r5a.4xlarge"|"r5a.8xlarge"|"r5a.12xlarge"|"r5a.16xlarge"|"r5a.24xlarge"|"m3.medium"|"m3.large"|"m3.xlarge"|"m3.2xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge"|"m5a.large"|"m5a.xlarge"|"m5a.2xlarge"|"m5a.4xlarge"|"m5a.8xlarge"|"m5a.12xlarge"|"m5a.16xlarge"|"m5a.24xlarge",
  EC2InboundPermissions = list(
    list(
      FromPort = 123,
      ToPort = 123,
      IpRange = "string",
      Protocol = "TCP"|"UDP"
    )
  ),
  NewGameSessionProtectionPolicy = "NoProtection"|"FullProtection",
  RuntimeConfiguration = list(
    ServerProcesses = list(
      list(
        LaunchPath = "string",
        Parameters = "string",
        ConcurrentExecutions = 123
      )
    ),
    MaxConcurrentGameSessionActivations = 123,
    GameSessionActivationTimeoutSeconds = 123
  ),
  ResourceCreationLimitPolicy = list(
    NewGameSessionsPerCreator = 123,
    PolicyPeriodInMinutes = 123
  ),
  MetricGroups = list(
    "string"
  ),
  PeerVpcAwsAccountId = "string",
  PeerVpcId = "string",
  FleetType = "ON_DEMAND"|"SPOT",
  InstanceRoleArn = "string",
  CertificateConfiguration = list(
    CertificateType = "DISABLED"|"GENERATED"
  ),
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
