% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_game_session}
\alias{gamelift_create_game_session}
\title{Creates a multiplayer game session for players}
\usage{
gamelift_create_game_session(FleetId, AliasId,
  MaximumPlayerSessionCount, Name, GameProperties, CreatorId,
  GameSessionId, IdempotencyToken, GameSessionData)
}
\arguments{
\item{FleetId}{A unique identifier for a fleet to create a game session in. You can use
either the fleet ID or ARN value. Each request must reference either a
fleet ID or alias ID, but not both.}

\item{AliasId}{A unique identifier for an alias associated with the fleet to create a
game session in. You can use either the alias ID or ARN value. Each
request must reference either a fleet ID or alias ID, but not both.}

\item{MaximumPlayerSessionCount}{[required] The maximum number of players that can be connected simultaneously to
the game session.}

\item{Name}{A descriptive label that is associated with a game session. Session
names do not need to be unique.}

\item{GameProperties}{Set of custom properties for a game session, formatted as key:value
pairs. These properties are passed to a game server process in the
GameSession object with a request to start a new game session (see
\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).}

\item{CreatorId}{A unique identifier for a player or entity creating the game session.
This ID is used to enforce a resource protection policy (if one exists)
that limits the number of concurrent active game sessions one player can
have.}

\item{GameSessionId}{\emph{This parameter is no longer preferred. Please use \code{IdempotencyToken}
instead.} Custom string that uniquely identifies a request for a new
game session. Maximum token length is 48 characters. If provided, this
string is included in the new game session's ID. (A game session ARN has
the following format:
\verb{arn:aws:gamelift:<region>::gamesession/<fleet ID>/<custom ID string or idempotency token>}.)}

\item{IdempotencyToken}{Custom string that uniquely identifies a request for a new game session.
Maximum token length is 48 characters. If provided, this string is
included in the new game session's ID. (A game session ARN has the
following format:
\verb{arn:aws:gamelift:<region>::gamesession/<fleet ID>/<custom ID string or idempotency token>}.)
Idempotency tokens remain in use for 30 days after a game session has
ended; game session objects are retained for this time period and then
deleted.}

\item{GameSessionData}{Set of custom game session properties, formatted as a single string
value. This data is passed to a game server process in the GameSession
object with a request to start a new game session (see \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-sdk-server-api.html#gamelift-sdk-server-startsession}{Start a Game Session}).}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GameSession = list(
    GameSessionId = "string",
    Name = "string",
    FleetId = "string",
    FleetArn = "string",
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    TerminationTime = as.POSIXct(
      "2015-01-01"
    ),
    CurrentPlayerSessionCount = 123,
    MaximumPlayerSessionCount = 123,
    Status = "ACTIVE"|"ACTIVATING"|"TERMINATED"|"TERMINATING"|"ERROR",
    StatusReason = "INTERRUPTED",
    GameProperties = list(
      list(
        Key = "string",
        Value = "string"
      )
    ),
    IpAddress = "string",
    DnsName = "string",
    Port = 123,
    PlayerSessionCreationPolicy = "ACCEPT_ALL"|"DENY_ALL",
    CreatorId = "string",
    GameSessionData = "string",
    MatchmakerData = "string"
  )
)
}\if{html}{\out{</div>}}
}
\description{
Creates a multiplayer game session for players. This operation creates a
game session record and assigns an available server process in the
specified fleet to host the game session. A fleet must have an \code{ACTIVE}
status before a game session can be created in it.

To create a game session, specify either fleet ID or alias ID and
indicate a maximum number of players to allow in the game session. You
can also provide a name and game-specific properties for this game
session. If successful, a GameSession object is returned containing the
game session properties and other settings you specified.

\strong{Idempotency tokens.} You can add a token that uniquely identifies
game session requests. This is useful for ensuring that game session
requests are idempotent. Multiple requests with the same idempotency
token are processed only once; subsequent requests return the original
result. All response values are the same with the exception of game
session status, which may change.

\strong{Resource creation limits.} If you are creating a game session on a
fleet with a resource creation limit policy in force, then you must
specify a creator ID. Without this ID, Amazon GameLift has no way to
evaluate the policy for this new game session request.

\strong{Player acceptance policy.} By default, newly created game sessions
are open to new players. You can restrict new player access by using
\code{\link[=gamelift_update_game_session]{update_game_session}} to change the game
session's player session creation policy.

\strong{Game session logs.} Logs are retained for all active game sessions
for 14 days. To access the logs, call
\code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}} to
download the log files.

\emph{Available in Amazon GameLift Local.}
\itemize{
\item \code{\link[=gamelift_create_game_session]{create_game_session}}
\item \code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}}
\item \code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
\item \code{\link[=gamelift_search_game_sessions]{search_game_sessions}}
\item \code{\link[=gamelift_update_game_session]{update_game_session}}
\item \code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}}
\item Game session placements
\itemize{
\item \code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
\item \code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
\item \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_game_session(
  FleetId = "string",
  AliasId = "string",
  MaximumPlayerSessionCount = 123,
  Name = "string",
  GameProperties = list(
    list(
      Key = "string",
      Value = "string"
    )
  ),
  CreatorId = "string",
  GameSessionId = "string",
  IdempotencyToken = "string",
  GameSessionData = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
