% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_matchmaking_rule_set}
\alias{gamelift_create_matchmaking_rule_set}
\title{Creates a new rule set for FlexMatch matchmaking}
\usage{
gamelift_create_matchmaking_rule_set(Name, RuleSetBody, Tags)
}
\arguments{
\item{Name}{[required] A unique identifier for a matchmaking rule set. A matchmaking
configuration identifies the rule set it uses by this name value. Note
that the rule set name is different from the optional \code{name} field in
the rule set body.}

\item{RuleSetBody}{[required] A collection of matchmaking rules, formatted as a JSON string. Comments
are not allowed in JSON, but most elements support a description field.}

\item{Tags}{A list of labels to assign to the new matchmaking rule set resource.
Tags are developer-defined key-value pairs. Tagging AWS resources are
useful for resource management, access management and cost allocation.
For more information, see \href{https://docs.aws.amazon.com/general/latest/gr/aws_tagging.html}{Tagging AWS Resources}
in the \emph{AWS General Reference}. Once the resource is created, you can
use TagResource, UntagResource, and ListTagsForResource to add, remove,
and view tags. The maximum tag limit may be lower than stated. See the
AWS General Reference for actual tagging limits.}
}
\description{
Creates a new rule set for FlexMatch matchmaking. A rule set describes
the type of match to create, such as the number and size of teams. It
also sets the parameters for acceptable player matches, such as minimum
skill level or character type. A rule set is used by a
MatchmakingConfiguration.

To create a matchmaking rule set, provide unique rule set name and the
rule set body in JSON format. Rule sets must be defined in the same
Region as the matchmaking configuration they are used with.

Since matchmaking rule sets cannot be edited, it is a good idea to check
the rule set syntax using ValidateMatchmakingRuleSet before creating a
new rule set.

\strong{Learn more}
\itemize{
\item \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-rulesets.html}{Build a Rule Set}
\item \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-configuration.html}{Design a Matchmaker}
\item \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-intro.html}{Matchmaking with FlexMatch}
}

\strong{Related operations}
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Request syntax}{
\preformatted{svc$create_matchmaking_rule_set(
  Name = "string",
  RuleSetBody = "string",
  Tags = list(
    list(
      Key = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
