% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_create_player_sessions}
\alias{gamelift_create_player_sessions}
\title{Reserves open slots in a game session for a group of players}
\usage{
gamelift_create_player_sessions(GameSessionId, PlayerIds, PlayerDataMap)
}
\arguments{
\item{GameSessionId}{[required] A unique identifier for the game session to add players to.}

\item{PlayerIds}{[required] List of unique identifiers for the players to be added.}

\item{PlayerDataMap}{Map of string pairs, each specifying a player ID and a set of
developer-defined information related to the player. Amazon GameLift
does not use this data, so it can be formatted as needed for use in the
game. Player data strings for player IDs not included in the \code{PlayerIds}
parameter are ignored.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  PlayerSessions = list(
    list(
      PlayerSessionId = "string",
      PlayerId = "string",
      GameSessionId = "string",
      FleetId = "string",
      FleetArn = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      TerminationTime = as.POSIXct(
        "2015-01-01"
      ),
      Status = "RESERVED"|"ACTIVE"|"COMPLETED"|"TIMEDOUT",
      IpAddress = "string",
      DnsName = "string",
      Port = 123,
      PlayerData = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Reserves open slots in a game session for a group of players. Before
players can be added, a game session must have an \code{ACTIVE} status, have
a creation policy of \code{ALLOW_ALL}, and have an open player slot. To add a
single player to a game session, use
\code{\link[=gamelift_create_player_session]{create_player_session}}. When a player
connects to the game server and references a player session ID, the game
server contacts the Amazon GameLift service to validate the player
reservation and accept the player.

To create player sessions, specify a game session ID, a list of player
IDs, and optionally a set of player data strings. If successful, a slot
is reserved in the game session for each player and a set of new
PlayerSession objects is returned. Player sessions cannot be updated.

\emph{Available in Amazon GameLift Local.}
\itemize{
\item \code{\link[=gamelift_create_player_session]{create_player_session}}
\item \code{\link[=gamelift_create_player_sessions]{create_player_sessions}}
\item \code{\link[=gamelift_describe_player_sessions]{describe_player_sessions}}
\item Game session placements
\itemize{
\item \code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
\item \code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
\item \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$create_player_sessions(
  GameSessionId = "string",
  PlayerIds = list(
    "string"
  ),
  PlayerDataMap = list(
    "string"
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
