% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_game_server_group}
\alias{gamelift_delete_game_server_group}
\title{This operation is used with the GameLift FleetIQ solution and game
server groups}
\usage{
gamelift_delete_game_server_group(GameServerGroupName, DeleteOption)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group. Use either the
GameServerGroup name or ARN value.}

\item{DeleteOption}{The type of delete to perform. Options include the following:
\itemize{
\item \code{SAFE_DELETE} – (default) Terminates the game server group and EC2
Auto Scaling group only when it has no game servers that are in
\code{UTILIZED} status.
\item \code{FORCE_DELETE} – Terminates the game server group, including all
active game servers regardless of their utilization status, and the
EC2 Auto Scaling group.
\item \code{RETAIN} – Does a safe delete of the game server group but retains
the EC2 Auto Scaling group as is.
}}
}
\value{
A list with the following syntax:\preformatted{list(
  GameServerGroup = list(
    GameServerGroupName = "string",
    GameServerGroupArn = "string",
    RoleArn = "string",
    InstanceDefinitions = list(
      list(
        InstanceType = "c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"c5a.large"|"c5a.xlarge"|"c5a.2xlarge"|"c5a.4xlarge"|"c5a.8xlarge"|"c5a.12xlarge"|"c5a.16xlarge"|"c5a.24xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"r5a.large"|"r5a.xlarge"|"r5a.2xlarge"|"r5a.4xlarge"|"r5a.8xlarge"|"r5a.12xlarge"|"r5a.16xlarge"|"r5a.24xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge"|"m5a.large"|"m5a.xlarge"|"m5a.2xlarge"|"m5a.4xlarge"|"m5a.8xlarge"|"m5a.12xlarge"|"m5a.16xlarge"|"m5a.24xlarge",
        WeightedCapacity = "string"
      )
    ),
    BalancingStrategy = "SPOT_ONLY"|"SPOT_PREFERRED"|"ON_DEMAND_ONLY",
    GameServerProtectionPolicy = "NO_PROTECTION"|"FULL_PROTECTION",
    AutoScalingGroupArn = "string",
    Status = "NEW"|"ACTIVATING"|"ACTIVE"|"DELETE_SCHEDULED"|"DELETING"|"DELETED"|"ERROR",
    StatusReason = "string",
    SuspendedActions = list(
      "REPLACE_INSTANCE_TYPES"
    ),
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
\strong{This operation is used with the GameLift FleetIQ solution and game
server groups.}

Terminates a game server group and permanently deletes the game server
group record. You have several options for how these resources are
impacted when deleting the game server group. Depending on the type of
delete operation selected, this operation might affect these resources:
\itemize{
\item The game server group
\item The corresponding Auto Scaling group
\item All game servers that are currently running in the group
}

To delete a game server group, identify the game server group to delete
and specify the type of delete operation to initiate. Game server groups
can only be deleted if they are in \code{ACTIVE} or \code{ERROR} status.

If the delete request is successful, a series of operations are kicked
off. The game server group status is changed to \code{DELETE_SCHEDULED},
which prevents new game servers from being registered and stops
automatic scaling activity. Once all game servers in the game server
group are deregistered, GameLift FleetIQ can begin deleting resources.
If any of the delete operations fail, the game server group is placed in
\code{ERROR} status.

GameLift FleetIQ emits delete events to Amazon CloudWatch.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related actions}

\code{\link[=gamelift_create_game_server_group]{create_game_server_group}} |
\code{\link[=gamelift_list_game_server_groups]{list_game_server_groups}} |
\code{\link[=gamelift_describe_game_server_group]{describe_game_server_group}} |
\code{\link[=gamelift_update_game_server_group]{update_game_server_group}} |
\code{\link[=gamelift_delete_game_server_group]{delete_game_server_group}} |
\code{\link[=gamelift_resume_game_server_group]{resume_game_server_group}} |
\code{\link[=gamelift_suspend_game_server_group]{suspend_game_server_group}} |
\code{\link[=gamelift_describe_game_server_instances]{describe_game_server_instances}}
| \href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/reference-awssdk-fleetiq.html}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$delete_game_server_group(
  GameServerGroupName = "string",
  DeleteOption = "SAFE_DELETE"|"FORCE_DELETE"|"RETAIN"
)
}
}

\keyword{internal}
