% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_matchmaking_configuration}
\alias{gamelift_delete_matchmaking_configuration}
\title{Permanently removes a FlexMatch matchmaking configuration}
\usage{
gamelift_delete_matchmaking_configuration(Name)
}
\arguments{
\item{Name}{[required] A unique identifier for the matchmaking configuration. You can use
either the configuration name or ARN value.}
}
\value{
An empty list.
}
\description{
Permanently removes a FlexMatch matchmaking configuration. To delete,
specify the configuration name. A matchmaking configuration cannot be
deleted if it is being used in any active matchmaking tickets.

\strong{Related actions}

\code{\link[=gamelift_create_matchmaking_configuration]{create_matchmaking_configuration}}
|
\code{\link[=gamelift_describe_matchmaking_configurations]{describe_matchmaking_configurations}}
|
\code{\link[=gamelift_update_matchmaking_configuration]{update_matchmaking_configuration}}
|
\code{\link[=gamelift_delete_matchmaking_configuration]{delete_matchmaking_configuration}}
| \code{\link[=gamelift_create_matchmaking_rule_set]{create_matchmaking_rule_set}}
|
\code{\link[=gamelift_describe_matchmaking_rule_sets]{describe_matchmaking_rule_sets}}
|
\code{\link[=gamelift_validate_matchmaking_rule_set]{validate_matchmaking_rule_set}}
| \code{\link[=gamelift_delete_matchmaking_rule_set]{delete_matchmaking_rule_set}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$delete_matchmaking_configuration(
  Name = "string"
)
}
}

\keyword{internal}
