% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_matchmaking_configuration}
\alias{gamelift_delete_matchmaking_configuration}
\title{Permanently removes a FlexMatch matchmaking configuration}
\usage{
gamelift_delete_matchmaking_configuration(Name)
}
\arguments{
\item{Name}{[required] Unique identifier for a matchmaking configuration}
}
\description{
Permanently removes a FlexMatch matchmaking configuration. To delete,
specify the configuration name. A matchmaking configuration cannot be
deleted if it is being used in any active matchmaking tickets.
}
\details{
\strong{Related operations}
\itemize{
\item CreateMatchmakingConfiguration
\item DescribeMatchmakingConfigurations
\item UpdateMatchmakingConfiguration
\item DeleteMatchmakingConfiguration
\item CreateMatchmakingRuleSet
\item DescribeMatchmakingRuleSets
\item ValidateMatchmakingRuleSet
\item DeleteMatchmakingRuleSet
}
}
\section{Request syntax}{
\preformatted{svc$delete_matchmaking_configuration(
  Name = "string"
)
}
}

\keyword{internal}
