% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_scaling_policy}
\alias{gamelift_delete_scaling_policy}
\title{Deletes a fleet scaling policy}
\usage{
gamelift_delete_scaling_policy(Name, FleetId)
}
\arguments{
\item{Name}{[required] A descriptive label that is associated with a scaling policy. Policy
names do not need to be unique.}

\item{FleetId}{[required] A unique identifier for a fleet to be deleted. You can use either the
fleet ID or ARN value.}
}
\value{
An empty list.
}
\description{
Deletes a fleet scaling policy. Once deleted, the policy is no longer in
force and GameLift removes all record of it. To delete a scaling policy,
specify both the scaling policy name and the fleet ID it is associated
with.

To temporarily suspend scaling policies, call
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}. This operation
suspends all policies for the fleet.
\itemize{
\item \code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}}
\item \code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}}
\item \code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
\item Manage scaling policies:
\itemize{
\item \code{\link[=gamelift_put_scaling_policy]{put_scaling_policy}}
(auto-scaling)
\item \code{\link[=gamelift_describe_scaling_policies]{describe_scaling_policies}}
(auto-scaling)
\item \code{\link[=gamelift_delete_scaling_policy]{delete_scaling_policy}}
(auto-scaling)
}
\item Manage fleet actions:
\itemize{
\item \code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}}
\item \code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}
}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$delete_scaling_policy(
  Name = "string",
  FleetId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
