% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_scaling_policy}
\alias{gamelift_delete_scaling_policy}
\title{Deletes a fleet scaling policy}
\usage{
gamelift_delete_scaling_policy(Name, FleetId)
}
\arguments{
\item{Name}{[required] Descriptive label that is associated with a scaling policy. Policy names
do not need to be unique.}

\item{FleetId}{[required] Unique identifier for a fleet to be deleted.}
}
\description{
Deletes a fleet scaling policy. This action means that the policy is no
longer in force and removes all record of it. To delete a scaling
policy, specify both the scaling policy name and the fleet ID it is
associated with.
}
\details{
To temporarily suspend scaling policies, call StopFleetActions. This
operation suspends all policies for the fleet.
\itemize{
\item DescribeFleetCapacity
\item UpdateFleetCapacity
\item DescribeEC2InstanceLimits
\item Manage scaling policies:
\item PutScalingPolicy (auto-scaling)
\item DescribeScalingPolicies (auto-scaling)
\item DeleteScalingPolicy (auto-scaling)
\item Manage fleet actions:
\item StartFleetActions
\item StopFleetActions
}
}
\section{Request syntax}{
\preformatted{svc$delete_scaling_policy(
  Name = "string",
  FleetId = "string"
)
}
}

\keyword{internal}
