% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_delete_vpc_peering_authorization}
\alias{gamelift_delete_vpc_peering_authorization}
\title{Cancels a pending VPC peering authorization for the specified VPC}
\usage{
gamelift_delete_vpc_peering_authorization(GameLiftAwsAccountId,
  PeerVpcId)
}
\arguments{
\item{GameLiftAwsAccountId}{[required] A unique identifier for the AWS account that you use to manage your
Amazon GameLift fleet. You can find your Account ID in the AWS
Management Console under account settings.}

\item{PeerVpcId}{[required] A unique identifier for a VPC with resources to be accessed by your
Amazon GameLift fleet. The VPC must be in the same Region where your
fleet is deployed. Look up a VPC ID using the \href{https://console.aws.amazon.com/vpc/home}{VPC Dashboard} in the AWS
Management Console. Learn more about VPC peering in \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/vpc-peering.html}{VPC Peering with Amazon GameLift Fleets}.}
}
\value{
A list with the following syntax:\preformatted{list()
}
}
\description{
Cancels a pending VPC peering authorization for the specified VPC. If
you need to delete an existing VPC peering connection, call
\code{\link[=gamelift_delete_vpc_peering_connection]{delete_vpc_peering_connection}}.
\itemize{
\item \code{\link[=gamelift_create_vpc_peering_authorization]{create_vpc_peering_authorization}}
\item \code{\link[=gamelift_describe_vpc_peering_authorizations]{describe_vpc_peering_authorizations}}
\item \code{\link[=gamelift_delete_vpc_peering_authorization]{delete_vpc_peering_authorization}}
\item \code{\link[=gamelift_create_vpc_peering_connection]{create_vpc_peering_connection}}
\item \code{\link[=gamelift_describe_vpc_peering_connections]{describe_vpc_peering_connections}}
\item \code{\link[=gamelift_delete_vpc_peering_connection]{delete_vpc_peering_connection}}
}
}
\section{Request syntax}{
\preformatted{svc$delete_vpc_peering_authorization(
  GameLiftAwsAccountId = "string",
  PeerVpcId = "string"
)
}
}

\keyword{internal}
