% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_deregister_game_server}
\alias{gamelift_deregister_game_server}
\title{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups}
\usage{
gamelift_deregister_game_server(GameServerGroupName, GameServerId)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group where the game server is
running. Use either the GameServerGroup name or ARN value.}

\item{GameServerId}{[required] A custom string that uniquely identifies the game server to deregister.}
}
\value{
An empty list.
}
\description{
\strong{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups.}

Removes the game server from a game server group. As a result of this
operation, the deregistered game server can no longer be claimed and
will not be returned in a list of active game servers.

To deregister a game server, specify the game server group and game
server ID. If successful, this operation emits a CloudWatch event with
termination timestamp and reason.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_register_game_server]{register_game_server}}
\item \code{\link[=gamelift_list_game_servers]{list_game_servers}}
\item \code{\link[=gamelift_claim_game_server]{claim_game_server}}
\item \code{\link[=gamelift_describe_game_server]{describe_game_server}}
\item \code{\link[=gamelift_update_game_server]{update_game_server}}
\item \code{\link[=gamelift_deregister_game_server]{deregister_game_server}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$deregister_game_server(
  GameServerGroupName = "string",
  GameServerId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
