% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_build}
\alias{gamelift_describe_build}
\title{Retrieves properties for a custom game build}
\usage{
gamelift_describe_build(BuildId)
}
\arguments{
\item{BuildId}{[required] A unique identifier for a build to retrieve properties for. You can use
either the build ID or ARN value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Build = list(
    BuildId = "string",
    BuildArn = "string",
    Name = "string",
    Version = "string",
    Status = "INITIALIZED"|"READY"|"FAILED",
    SizeOnDisk = 123,
    OperatingSystem = "WINDOWS_2012"|"AMAZON_LINUX"|"AMAZON_LINUX_2",
    CreationTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves properties for a custom game build. To request a build
resource, specify a build ID. If successful, an object containing the
build properties is returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-build-intro.html}{Upload a Custom Server Build}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_build]{create_build}}
\item \code{\link[=gamelift_list_builds]{list_builds}}
\item \code{\link[=gamelift_describe_build]{describe_build}}
\item \code{\link[=gamelift_update_build]{update_build}}
\item \code{\link[=gamelift_delete_build]{delete_build}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_build(
  BuildId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
