% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_port_settings}
\alias{gamelift_describe_fleet_port_settings}
\title{Retrieves a fleet's inbound connection permissions}
\usage{
gamelift_describe_fleet_port_settings(FleetId)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet to retrieve port settings for. You can
use either the fleet ID or ARN value.}
}
\value{
A list with the following syntax:\preformatted{list(
  InboundPermissions = list(
    list(
      FromPort = 123,
      ToPort = 123,
      IpRange = "string",
      Protocol = "TCP"|"UDP"
    )
  )
)
}
}
\description{
Retrieves a fleet's inbound connection permissions. Connection
permissions specify the range of IP addresses and port settings that
incoming traffic can use to access server processes in the fleet. Game
sessions that are running on instances in the fleet use connections that
fall in this range.

To get a fleet's inbound connection permissions, specify the fleet's
unique identifier. If successful, a collection of IpPermission objects
is returned for the requested fleet ID. If the requested fleet has been
deleted, the result set is empty.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_fleet]{create_fleet}}
\item \code{\link[=gamelift_list_fleets]{list_fleets}}
\item \code{\link[=gamelift_delete_fleet]{delete_fleet}}
\item Describe fleets:
\itemize{
\item \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}}
\item \code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}}
\item \code{\link[=gamelift_describe_fleet_port_settings]{describe_fleet_port_settings}}
\item \code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}}
\item \code{\link[=gamelift_describe_runtime_configuration]{describe_runtime_configuration}}
\item \code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
\item \code{\link[=gamelift_describe_fleet_events]{describe_fleet_events}}
}
\item \code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}}
\item \code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} or
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}
}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_port_settings(
  FleetId = "string"
)
}
}

\keyword{internal}
