% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_port_settings}
\alias{gamelift_describe_fleet_port_settings}
\title{Retrieves the inbound connection permissions for a fleet}
\usage{
gamelift_describe_fleet_port_settings(FleetId)
}
\arguments{
\item{FleetId}{[required] Unique identifier for a fleet to retrieve port settings for.}
}
\description{
Retrieves the inbound connection permissions for a fleet. Connection
permissions include a range of IP addresses and port settings that
incoming traffic can use to access server processes in the fleet. To get
a fleet\'s inbound connection permissions, specify a fleet ID. If
successful, a collection of IpPermission objects is returned for the
requested fleet ID. If the requested fleet has been deleted, the result
set is empty.
}
\details{
\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Working with Fleets}.

\strong{Related operations}
\itemize{
\item CreateFleet
\item ListFleets
\item DeleteFleet
\item Describe fleets:
\itemize{
\item DescribeFleetAttributes
\item DescribeFleetCapacity
\item DescribeFleetPortSettings
\item DescribeFleetUtilization
\item DescribeRuntimeConfiguration
\item DescribeEC2InstanceLimits
\item DescribeFleetEvents
}
\item Update fleets:
\itemize{
\item UpdateFleetAttributes
\item UpdateFleetCapacity
\item UpdateFleetPortSettings
\item UpdateRuntimeConfiguration
}
\item Manage fleet actions:
\itemize{
\item StartFleetActions
\item StopFleetActions
}
}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_port_settings(
  FleetId = "string"
)
}
}

\keyword{internal}
