% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_fleet_utilization}
\alias{gamelift_describe_fleet_utilization}
\title{Retrieves utilization statistics for one or more fleets}
\usage{
gamelift_describe_fleet_utilization(FleetIds, Limit, NextToken)
}
\arguments{
\item{FleetIds}{A unique identifier for a fleet(s) to retrieve utilization data for. You
can use either the fleet ID or ARN value. To retrieve attributes for all
current fleets, do not include this parameter. If the list of fleet
identifiers includes fleets that don't currently exist, the request
succeeds but no attributes for that fleet are returned.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages. This parameter
is ignored when the request specifies one or a list of fleet IDs.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.
This parameter is ignored when the request specifies one or a list of
fleet IDs.}
}
\value{
A list with the following syntax:\preformatted{list(
  FleetUtilization = list(
    list(
      FleetId = "string",
      ActiveServerProcessCount = 123,
      ActiveGameSessionCount = 123,
      CurrentPlayerSessionCount = 123,
      MaximumPlayerSessionCount = 123
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves utilization statistics for one or more fleets. These
statistics provide insight into how available hosting resources are
currently being used. To get statistics on available hosting resources,
see \code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}}.

You can request utilization data for all fleets, or specify a list of
one or more fleet IDs. When requesting multiple fleets, use the
pagination parameters to retrieve results as a set of sequential pages.
If successful, a FleetUtilization object is returned for each requested
fleet ID, unless the fleet identifier is not found.

Some API operations may limit the number of fleet IDs allowed in one
request. If a request exceeds this limit, the request fails and the
error message includes the maximum allowed.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/monitoring-cloudwatch.html#gamelift-metrics-fleet}{GameLift Metrics for Fleets}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_fleet]{create_fleet}}
\item \code{\link[=gamelift_list_fleets]{list_fleets}}
\item \code{\link[=gamelift_delete_fleet]{delete_fleet}}
\item Describe fleets:
\itemize{
\item \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}}
\item \code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}}
\item \code{\link[=gamelift_describe_fleet_port_settings]{describe_fleet_port_settings}}
\item \code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}}
\item \code{\link[=gamelift_describe_runtime_configuration]{describe_runtime_configuration}}
\item \code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
\item \code{\link[=gamelift_describe_fleet_events]{describe_fleet_events}}
}
\item \code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}}
\item \code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} or
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}
}
}
\section{Request syntax}{
\preformatted{svc$describe_fleet_utilization(
  FleetIds = list(
    "string"
  ),
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
