% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_game_server}
\alias{gamelift_describe_game_server}
\title{This operation is used with the GameLift FleetIQ solution and game
server groups}
\usage{
gamelift_describe_game_server(GameServerGroupName, GameServerId)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group where the game server is
running. Use either the GameServerGroup name or ARN value.}

\item{GameServerId}{[required] A custom string that uniquely identifies the game server information to
be retrieved.}
}
\value{
A list with the following syntax:\preformatted{list(
  GameServer = list(
    GameServerGroupName = "string",
    GameServerGroupArn = "string",
    GameServerId = "string",
    InstanceId = "string",
    ConnectionInfo = "string",
    GameServerData = "string",
    ClaimStatus = "CLAIMED",
    UtilizationStatus = "AVAILABLE"|"UTILIZED",
    RegistrationTime = as.POSIXct(
      "2015-01-01"
    ),
    LastClaimTime = as.POSIXct(
      "2015-01-01"
    ),
    LastHealthCheckTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
\strong{This operation is used with the GameLift FleetIQ solution and game
server groups.}

Retrieves information for a registered game server. Information includes
game server status, health check info, and the instance that the game
server is running on.

To retrieve game server information, specify the game server ID. If
successful, the requested game server object is returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related actions}

\code{\link[=gamelift_register_game_server]{register_game_server}} |
\code{\link[=gamelift_list_game_servers]{list_game_servers}} |
\code{\link[=gamelift_claim_game_server]{claim_game_server}} |
\code{\link[=gamelift_describe_game_server]{describe_game_server}} |
\code{\link[=gamelift_update_game_server]{update_game_server}} |
\code{\link[=gamelift_deregister_game_server]{deregister_game_server}} | \href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/reference-awssdk-fleetiq.html}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_game_server(
  GameServerGroupName = "string",
  GameServerId = "string"
)
}
}

\keyword{internal}
