% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_game_server_group}
\alias{gamelift_describe_game_server_group}
\title{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups}
\usage{
gamelift_describe_game_server_group(GameServerGroupName)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group. Use either the
GameServerGroup name or ARN value.}
}
\description{
\strong{This operation is used with the Amazon GameLift FleetIQ solution and
game server groups.}

Retrieves information on a game server group. This operation returns
only properties related to GameLift FleetIQ. To view or update
properties for the corresponding Auto Scaling group, such as launch
template, auto scaling policies, and maximum/minimum group size, access
the Auto Scaling group directly.

To get attributes for a game server group, provide a group name or ARN
value. If successful, a GameServerGroup object is returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_game_server_group]{create_game_server_group}}
\item \code{\link[=gamelift_list_game_server_groups]{list_game_server_groups}}
\item \code{\link[=gamelift_describe_game_server_group]{describe_game_server_group}}
\item \code{\link[=gamelift_update_game_server_group]{update_game_server_group}}
\item \code{\link[=gamelift_delete_game_server_group]{delete_game_server_group}}
\item \code{\link[=gamelift_resume_game_server_group]{resume_game_server_group}}
\item \code{\link[=gamelift_suspend_game_server_group]{suspend_game_server_group}}
\item \code{\link[=gamelift_describe_game_server_instances]{describe_game_server_instances}}
}
}
\section{Request syntax}{
\preformatted{svc$describe_game_server_group(
  GameServerGroupName = "string"
)
}
}

\keyword{internal}
