% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_game_session_details}
\alias{gamelift_describe_game_session_details}
\title{Retrieves properties, including the protection policy in force, for one
or more game sessions}
\usage{
gamelift_describe_game_session_details(FleetId, GameSessionId, AliasId,
  StatusFilter, Limit, NextToken)
}
\arguments{
\item{FleetId}{A unique identifier for a fleet to retrieve all game sessions active on
the fleet. You can use either the fleet ID or ARN value.}

\item{GameSessionId}{A unique identifier for the game session to retrieve.}

\item{AliasId}{A unique identifier for an alias associated with the fleet to retrieve
all game sessions for. You can use either the alias ID or ARN value.}

\item{StatusFilter}{Game session status to filter results on. Possible game session statuses
include \code{ACTIVE}, \code{TERMINATED}, \code{ACTIVATING} and \code{TERMINATING} (the last
two are transitory).}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  GameSessionDetails = list(
    list(
      GameSession = list(
        GameSessionId = "string",
        Name = "string",
        FleetId = "string",
        FleetArn = "string",
        CreationTime = as.POSIXct(
          "2015-01-01"
        ),
        TerminationTime = as.POSIXct(
          "2015-01-01"
        ),
        CurrentPlayerSessionCount = 123,
        MaximumPlayerSessionCount = 123,
        Status = "ACTIVE"|"ACTIVATING"|"TERMINATED"|"TERMINATING"|"ERROR",
        StatusReason = "INTERRUPTED",
        GameProperties = list(
          list(
            Key = "string",
            Value = "string"
          )
        ),
        IpAddress = "string",
        DnsName = "string",
        Port = 123,
        PlayerSessionCreationPolicy = "ACCEPT_ALL"|"DENY_ALL",
        CreatorId = "string",
        GameSessionData = "string",
        MatchmakerData = "string"
      ),
      ProtectionPolicy = "NoProtection"|"FullProtection"
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves properties, including the protection policy in force, for one
or more game sessions. This operation can be used in several ways: (1)
provide a \code{GameSessionId} or \code{GameSessionArn} to request details for a
specific game session; (2) provide either a \code{FleetId} or an \code{AliasId} to
request properties for all game sessions running on a fleet.

To get game session record(s), specify just one of the following: game
session ID, fleet ID, or alias ID. You can filter this request by game
session status. Use the pagination parameters to retrieve results as a
set of sequential pages. If successful, a GameSessionDetail object is
returned for each session matching the request.
\itemize{
\item \code{\link[=gamelift_create_game_session]{create_game_session}}
\item \code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}}
\item \code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
\item \code{\link[=gamelift_search_game_sessions]{search_game_sessions}}
\item \code{\link[=gamelift_update_game_session]{update_game_session}}
\item \code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}}
\item Game session placements
\itemize{
\item \code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
\item \code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
\item \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_game_session_details(
  FleetId = "string",
  GameSessionId = "string",
  AliasId = "string",
  StatusFilter = "string",
  Limit = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
