% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_instances}
\alias{gamelift_describe_instances}
\title{Retrieves information about a fleet's instances, including instance IDs,
connection data, and status}
\usage{
gamelift_describe_instances(FleetId, InstanceId, Limit, NextToken,
  Location)
}
\arguments{
\item{FleetId}{[required] A unique identifier for the fleet to retrieve instance information for.
You can use either the fleet ID or ARN value.}

\item{InstanceId}{A unique identifier for an instance to retrieve. Specify an instance ID
or leave blank to retrieve all instances in the fleet.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}

\item{Location}{The name of a location to retrieve instance information for, in the form
of an AWS Region code such as \code{us-west-2}.}
}
\value{
A list with the following syntax:\preformatted{list(
  Instances = list(
    list(
      FleetId = "string",
      FleetArn = "string",
      InstanceId = "string",
      IpAddress = "string",
      DnsName = "string",
      OperatingSystem = "WINDOWS_2012"|"AMAZON_LINUX"|"AMAZON_LINUX_2",
      Type = "t2.micro"|"t2.small"|"t2.medium"|"t2.large"|"c3.large"|"c3.xlarge"|"c3.2xlarge"|"c3.4xlarge"|"c3.8xlarge"|"c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"c5a.large"|"c5a.xlarge"|"c5a.2xlarge"|"c5a.4xlarge"|"c5a.8xlarge"|"c5a.12xlarge"|"c5a.16xlarge"|"c5a.24xlarge"|"r3.large"|"r3.xlarge"|"r3.2xlarge"|"r3.4xlarge"|"r3.8xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"r5a.large"|"r5a.xlarge"|"r5a.2xlarge"|"r5a.4xlarge"|"r5a.8xlarge"|"r5a.12xlarge"|"r5a.16xlarge"|"r5a.24xlarge"|"m3.medium"|"m3.large"|"m3.xlarge"|"m3.2xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge"|"m5a.large"|"m5a.xlarge"|"m5a.2xlarge"|"m5a.4xlarge"|"m5a.8xlarge"|"m5a.12xlarge"|"m5a.16xlarge"|"m5a.24xlarge",
      Status = "PENDING"|"ACTIVE"|"TERMINATING",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      Location = "string"
    )
  ),
  NextToken = "string"
)
}
}
\description{
Retrieves information about a fleet's instances, including instance IDs,
connection data, and status.

This operation can be used in the following ways:
\itemize{
\item To get information on all instances that are deployed to a fleet's
home Region, provide the fleet ID.
\item To get information on all instances that are deployed to a fleet's
remote location, provide the fleet ID and location name.
\item To get information on a specific instance in a fleet, provide the
fleet ID and instance ID.
}

Use the pagination parameters to retrieve results as a set of sequential
pages.

If successful, an \code{Instance} object is returned for each requested
instance. Instances are not returned in any particular order.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-remote-access.html}{Remotely Access Fleet Instances}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-creating-debug.html}{Debug Fleet Issues}

\strong{Related actions}

\code{\link[=gamelift_describe_instances]{describe_instances}} |
\code{\link[=gamelift_get_instance_access]{get_instance_access}} |
\code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_instances(
  FleetId = "string",
  InstanceId = "string",
  Limit = 123,
  NextToken = "string",
  Location = "string"
)
}
}

\keyword{internal}
