% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_runtime_configuration}
\alias{gamelift_describe_runtime_configuration}
\title{Retrieves a fleet's runtime configuration settings}
\usage{
gamelift_describe_runtime_configuration(FleetId)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet to get the runtime configuration for.
You can use either the fleet ID or ARN value.}
}
\value{
A list with the following syntax:\preformatted{list(
  RuntimeConfiguration = list(
    ServerProcesses = list(
      list(
        LaunchPath = "string",
        Parameters = "string",
        ConcurrentExecutions = 123
      )
    ),
    MaxConcurrentGameSessionActivations = 123,
    GameSessionActivationTimeoutSeconds = 123
  )
)
}
}
\description{
Retrieves a fleet's runtime configuration settings. The runtime
configuration tells Amazon GameLift which server processes to run (and
how) on each instance in the fleet.

To get a runtime configuration, specify the fleet's unique identifier.
If successful, a RuntimeConfiguration object is returned for the
requested fleet. If the requested fleet has been deleted, the result set
is empty.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-multiprocess.html}{Running Multiple Processes on a Fleet}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_fleet]{create_fleet}}
\item \code{\link[=gamelift_list_fleets]{list_fleets}}
\item \code{\link[=gamelift_delete_fleet]{delete_fleet}}
\item Describe fleets:
\itemize{
\item \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}}
\item \code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}}
\item \code{\link[=gamelift_describe_fleet_port_settings]{describe_fleet_port_settings}}
\item \code{\link[=gamelift_describe_fleet_utilization]{describe_fleet_utilization}}
\item \code{\link[=gamelift_describe_runtime_configuration]{describe_runtime_configuration}}
\item \code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
\item \code{\link[=gamelift_describe_fleet_events]{describe_fleet_events}}
}
\item \code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}}
\item \code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} or
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}
}
}
\section{Request syntax}{
\preformatted{svc$describe_runtime_configuration(
  FleetId = "string"
)
}
}

\keyword{internal}
