% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_vpc_peering_authorizations}
\alias{gamelift_describe_vpc_peering_authorizations}
\title{Retrieves valid VPC peering authorizations that are pending for the AWS
account}
\usage{
gamelift_describe_vpc_peering_authorizations()
}
\value{
A list with the following syntax:\preformatted{list(
  VpcPeeringAuthorizations = list(
    list(
      GameLiftAwsAccountId = "string",
      PeerVpcAwsAccountId = "string",
      PeerVpcId = "string",
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      ExpirationTime = as.POSIXct(
        "2015-01-01"
      )
    )
  )
)
}
}
\description{
Retrieves valid VPC peering authorizations that are pending for the AWS
account. This operation returns all VPC peering authorizations and
requests for peering. This includes those initiated and received by this
account.

\strong{Related actions}

\code{\link[=gamelift_create_vpc_peering_authorization]{create_vpc_peering_authorization}}
|
\code{\link[=gamelift_describe_vpc_peering_authorizations]{describe_vpc_peering_authorizations}}
|
\code{\link[=gamelift_delete_vpc_peering_authorization]{delete_vpc_peering_authorization}}
|
\code{\link[=gamelift_create_vpc_peering_connection]{create_vpc_peering_connection}}
|
\code{\link[=gamelift_describe_vpc_peering_connections]{describe_vpc_peering_connections}}
|
\code{\link[=gamelift_delete_vpc_peering_connection]{delete_vpc_peering_connection}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$describe_vpc_peering_authorizations()
}
}

\keyword{internal}
