% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_vpc_peering_connections}
\alias{gamelift_describe_vpc_peering_connections}
\title{Retrieves information on VPC peering connections}
\usage{
gamelift_describe_vpc_peering_connections(FleetId)
}
\arguments{
\item{FleetId}{A unique identifier for a fleet. You can use either the fleet ID or ARN
value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  VpcPeeringConnections = list(
    list(
      FleetId = "string",
      FleetArn = "string",
      IpV4CidrBlock = "string",
      VpcPeeringConnectionId = "string",
      Status = list(
        Code = "string",
        Message = "string"
      ),
      PeerVpcId = "string",
      GameLiftVpcId = "string"
    )
  )
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves information on VPC peering connections. Use this operation to
get peering information for all fleets or for one specific fleet ID.

To retrieve connection information, call this operation from the AWS
account that is used to manage the Amazon GameLift fleets. Specify a
fleet ID or leave the parameter empty to retrieve all connection
records. If successful, the retrieved information includes both active
and pending connections. Active connections identify the IpV4 CIDR block
that the VPC uses to connect.
\itemize{
\item \code{\link[=gamelift_create_vpc_peering_authorization]{create_vpc_peering_authorization}}
\item \code{\link[=gamelift_describe_vpc_peering_authorizations]{describe_vpc_peering_authorizations}}
\item \code{\link[=gamelift_delete_vpc_peering_authorization]{delete_vpc_peering_authorization}}
\item \code{\link[=gamelift_create_vpc_peering_connection]{create_vpc_peering_connection}}
\item \code{\link[=gamelift_describe_vpc_peering_connections]{describe_vpc_peering_connections}}
\item \code{\link[=gamelift_delete_vpc_peering_connection]{delete_vpc_peering_connection}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$describe_vpc_peering_connections(
  FleetId = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
