% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_get_game_session_log_url}
\alias{gamelift_get_game_session_log_url}
\title{Retrieves the location of stored game session logs for a specified game
session}
\usage{
gamelift_get_game_session_log_url(GameSessionId)
}
\arguments{
\item{GameSessionId}{[required] A unique identifier for the game session to get logs for.}
}
\value{
A list with the following syntax:\preformatted{list(
  PreSignedUrl = "string"
)
}
}
\description{
Retrieves the location of stored game session logs for a specified game
session. When a game session is terminated, GameLift automatically
stores the logs in Amazon S3 and retains them for 14 days. Use this URL
to download the logs.

See the \href{https://docs.aws.amazon.com/general/latest/gr/aws_service_limits.html#limits_gamelift}{AWS Service Limits}
page for maximum log file sizes. Log files that exceed this limit are
not saved.

\strong{Related actions}

\code{\link[=gamelift_create_game_session]{create_game_session}} |
\code{\link[=gamelift_describe_game_sessions]{describe_game_sessions}} |
\code{\link[=gamelift_describe_game_session_details]{describe_game_session_details}}
| \code{\link[=gamelift_search_game_sessions]{search_game_sessions}} |
\code{\link[=gamelift_update_game_session]{update_game_session}} |
\code{\link[=gamelift_get_game_session_log_url]{get_game_session_log_url}} |
\code{\link[=gamelift_start_game_session_placement]{start_game_session_placement}}
|
\code{\link[=gamelift_describe_game_session_placement]{describe_game_session_placement}}
| \code{\link[=gamelift_stop_game_session_placement]{stop_game_session_placement}}
| \href{https://docs.aws.amazon.com/gamelift/latest/developerguide/reference-awssdk.html#reference-awssdk-resources-fleets}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$get_game_session_log_url(
  GameSessionId = "string"
)
}
}

\keyword{internal}
