% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_fleets}
\alias{gamelift_list_fleets}
\title{Retrieves a collection of fleet resources for this AWS account}
\usage{
gamelift_list_fleets(BuildId, ScriptId, Limit, NextToken)
}
\arguments{
\item{BuildId}{A unique identifier for a build to return fleets for. Use this parameter
to return only fleets using a specified build. Use either the build ID
or ARN value. To retrieve all fleets, do not include either a BuildId
and ScriptID parameter.}

\item{ScriptId}{A unique identifier for a Realtime script to return fleets for. Use this
parameter to return only fleets using a specified script. Use either the
script ID or ARN value. To retrieve all fleets, leave this parameter
empty.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves a collection of fleet resources for this AWS account. You can
filter the result set to find only those fleets that are deployed with a
specific build or script. Use the pagination parameters to retrieve
results in sequential pages.

Fleet resources are not listed in a particular order.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_fleet]{create_fleet}}
\item \code{\link[=gamelift_list_fleets]{list_fleets}}
\item \code{\link[=gamelift_delete_fleet]{delete_fleet}}
\item \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}}
\item \code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}}
\item \code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} or
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}
}
}
\section{Request syntax}{
\preformatted{svc$list_fleets(
  BuildId = "string",
  ScriptId = "string",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
