% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_game_servers}
\alias{gamelift_list_game_servers}
\title{This operation is used with the GameLift FleetIQ solution and game
server groups}
\usage{
gamelift_list_game_servers(GameServerGroupName, SortOrder, Limit,
  NextToken)
}
\arguments{
\item{GameServerGroupName}{[required] An identifier for the game server group to retrieve a list of game
servers from. Use either the GameServerGroup name or ARN value.}

\item{SortOrder}{Indicates how to sort the returned data based on game server
registration timestamp. Use \code{ASCENDING} to retrieve oldest game servers
first, or use \code{DESCENDING} to retrieve newest game servers first. If
this parameter is left empty, game servers are returned in no particular
order.}

\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\value{
A list with the following syntax:\preformatted{list(
  GameServers = list(
    list(
      GameServerGroupName = "string",
      GameServerGroupArn = "string",
      GameServerId = "string",
      InstanceId = "string",
      ConnectionInfo = "string",
      GameServerData = "string",
      ClaimStatus = "CLAIMED",
      UtilizationStatus = "AVAILABLE"|"UTILIZED",
      RegistrationTime = as.POSIXct(
        "2015-01-01"
      ),
      LastClaimTime = as.POSIXct(
        "2015-01-01"
      ),
      LastHealthCheckTime = as.POSIXct(
        "2015-01-01"
      )
    )
  ),
  NextToken = "string"
)
}
}
\description{
\strong{This operation is used with the GameLift FleetIQ solution and game
server groups.}

Retrieves information on all game servers that are currently active in a
specified game server group. You can opt to sort the list by game server
age. Use the pagination parameters to retrieve results in a set of
sequential segments.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related actions}

\code{\link[=gamelift_register_game_server]{register_game_server}} |
\code{\link[=gamelift_list_game_servers]{list_game_servers}} |
\code{\link[=gamelift_claim_game_server]{claim_game_server}} |
\code{\link[=gamelift_describe_game_server]{describe_game_server}} |
\code{\link[=gamelift_update_game_server]{update_game_server}} |
\code{\link[=gamelift_deregister_game_server]{deregister_game_server}} | \href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/reference-awssdk-fleetiq.html}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$list_game_servers(
  GameServerGroupName = "string",
  SortOrder = "ASCENDING"|"DESCENDING",
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
