% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_scripts}
\alias{gamelift_list_scripts}
\title{Retrieves script records for all Realtime scripts that are associated
with the AWS account in use}
\usage{
gamelift_list_scripts(Limit, NextToken)
}
\arguments{
\item{Limit}{The maximum number of results to return. Use this parameter with
\code{NextToken} to get results as a set of sequential pages.}

\item{NextToken}{A token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this operation.
To start at the beginning of the result set, do not specify a value.}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  Scripts = list(
    list(
      ScriptId = "string",
      ScriptArn = "string",
      Name = "string",
      Version = "string",
      SizeOnDisk = 123,
      CreationTime = as.POSIXct(
        "2015-01-01"
      ),
      StorageLocation = list(
        Bucket = "string",
        Key = "string",
        RoleArn = "string",
        ObjectVersion = "string"
      )
    )
  ),
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}
\description{
Retrieves script records for all Realtime scripts that are associated
with the AWS account in use.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/realtime-intro.html}{Amazon GameLift Realtime Servers}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_script]{create_script}}
\item \code{\link[=gamelift_list_scripts]{list_scripts}}
\item \code{\link[=gamelift_describe_script]{describe_script}}
\item \code{\link[=gamelift_update_script]{update_script}}
\item \code{\link[=gamelift_delete_script]{delete_script}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$list_scripts(
  Limit = 123,
  NextToken = "string"
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
