% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_list_scripts}
\alias{gamelift_list_scripts}
\title{Retrieves script records for all Realtime scripts that are associated
with the AWS account in use}
\usage{
gamelift_list_scripts(Limit, NextToken)
}
\arguments{
\item{Limit}{Maximum number of results to return. Use this parameter with \code{NextToken}
to get results as a set of sequential pages.}

\item{NextToken}{Token that indicates the start of the next sequential page of results.
Use the token that is returned with a previous call to this action. To
start at the beginning of the result set, do not specify a value.}
}
\description{
Retrieves script records for all Realtime scripts that are associated
with the AWS account in use.
}
\details{
\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/realtime-intro.html}{Amazon GameLift Realtime Servers}

\strong{Related operations}
\itemize{
\item CreateScript
\item ListScripts
\item DescribeScript
\item UpdateScript
\item DeleteScript
}
}
\section{Request syntax}{
\preformatted{svc$list_scripts(
  Limit = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
