% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_resume_game_server_group}
\alias{gamelift_resume_game_server_group}
\title{This operation is used with the GameLift FleetIQ solution and game
server groups}
\usage{
gamelift_resume_game_server_group(GameServerGroupName, ResumeActions)
}
\arguments{
\item{GameServerGroupName}{[required] A unique identifier for the game server group. Use either the
GameServerGroup name or ARN value.}

\item{ResumeActions}{[required] The activity to resume for this game server group.}
}
\value{
A list with the following syntax:\preformatted{list(
  GameServerGroup = list(
    GameServerGroupName = "string",
    GameServerGroupArn = "string",
    RoleArn = "string",
    InstanceDefinitions = list(
      list(
        InstanceType = "c4.large"|"c4.xlarge"|"c4.2xlarge"|"c4.4xlarge"|"c4.8xlarge"|"c5.large"|"c5.xlarge"|"c5.2xlarge"|"c5.4xlarge"|"c5.9xlarge"|"c5.12xlarge"|"c5.18xlarge"|"c5.24xlarge"|"c5a.large"|"c5a.xlarge"|"c5a.2xlarge"|"c5a.4xlarge"|"c5a.8xlarge"|"c5a.12xlarge"|"c5a.16xlarge"|"c5a.24xlarge"|"r4.large"|"r4.xlarge"|"r4.2xlarge"|"r4.4xlarge"|"r4.8xlarge"|"r4.16xlarge"|"r5.large"|"r5.xlarge"|"r5.2xlarge"|"r5.4xlarge"|"r5.8xlarge"|"r5.12xlarge"|"r5.16xlarge"|"r5.24xlarge"|"r5a.large"|"r5a.xlarge"|"r5a.2xlarge"|"r5a.4xlarge"|"r5a.8xlarge"|"r5a.12xlarge"|"r5a.16xlarge"|"r5a.24xlarge"|"m4.large"|"m4.xlarge"|"m4.2xlarge"|"m4.4xlarge"|"m4.10xlarge"|"m5.large"|"m5.xlarge"|"m5.2xlarge"|"m5.4xlarge"|"m5.8xlarge"|"m5.12xlarge"|"m5.16xlarge"|"m5.24xlarge"|"m5a.large"|"m5a.xlarge"|"m5a.2xlarge"|"m5a.4xlarge"|"m5a.8xlarge"|"m5a.12xlarge"|"m5a.16xlarge"|"m5a.24xlarge",
        WeightedCapacity = "string"
      )
    ),
    BalancingStrategy = "SPOT_ONLY"|"SPOT_PREFERRED"|"ON_DEMAND_ONLY",
    GameServerProtectionPolicy = "NO_PROTECTION"|"FULL_PROTECTION",
    AutoScalingGroupArn = "string",
    Status = "NEW"|"ACTIVATING"|"ACTIVE"|"DELETE_SCHEDULED"|"DELETING"|"DELETED"|"ERROR",
    StatusReason = "string",
    SuspendedActions = list(
      "REPLACE_INSTANCE_TYPES"
    ),
    CreationTime = as.POSIXct(
      "2015-01-01"
    ),
    LastUpdatedTime = as.POSIXct(
      "2015-01-01"
    )
  )
)
}
}
\description{
\strong{This operation is used with the GameLift FleetIQ solution and game
server groups.}

Reinstates activity on a game server group after it has been suspended.
A game server group might be suspended by
the\code{\link[=gamelift_suspend_game_server_group]{suspend_game_server_group}}
operation, or it might be suspended involuntarily due to a configuration
problem. In the second case, you can manually resume activity on the
group once the configuration problem has been resolved. Refer to the
game server group status and status reason for more information on why
group activity is suspended.

To resume activity, specify a game server group ARN and the type of
activity to be resumed. If successful, a GameServerGroup object is
returned showing that the resumed activity is no longer listed in
\code{SuspendedActions}.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related actions}

\code{\link[=gamelift_create_game_server_group]{create_game_server_group}} |
\code{\link[=gamelift_list_game_server_groups]{list_game_server_groups}} |
\code{\link[=gamelift_describe_game_server_group]{describe_game_server_group}} |
\code{\link[=gamelift_update_game_server_group]{update_game_server_group}} |
\code{\link[=gamelift_delete_game_server_group]{delete_game_server_group}} |
\code{\link[=gamelift_resume_game_server_group]{resume_game_server_group}} |
\code{\link[=gamelift_suspend_game_server_group]{suspend_game_server_group}} |
\code{\link[=gamelift_describe_game_server_instances]{describe_game_server_instances}}
| \href{https://docs.aws.amazon.com/gamelift/latest/fleetiqguide/reference-awssdk-fleetiq.html}{All APIs by task}
}
\section{Request syntax}{
\preformatted{svc$resume_game_server_group(
  GameServerGroupName = "string",
  ResumeActions = list(
    "REPLACE_INSTANCE_TYPES"
  )
)
}
}

\keyword{internal}
