% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_resume_game_server_group}
\alias{gamelift_resume_game_server_group}
\title{This action is part of Amazon GameLift FleetIQ with game server groups,
which is in preview release and is subject to change}
\usage{
gamelift_resume_game_server_group(GameServerGroupName, ResumeActions)
}
\arguments{
\item{GameServerGroupName}{[required] The unique identifier of the game server group to resume activity on.
Use either the GameServerGroup name or ARN value.}

\item{ResumeActions}{[required] The action to resume for this game server group.}
}
\description{
\strong{This action is part of Amazon GameLift FleetIQ with game server
groups, which is in preview release and is subject to change.}
}
\details{
Reinstates activity on a game server group after it has been suspended.
A game server group may be suspended by calling SuspendGameServerGroup,
or it may have been involuntarily suspended due to a configuration
problem. You can manually resume activity on the group once the
configuration problem has been resolved. Refer to the game server group
status and status reason for more information on why group activity is
suspended.

To resume activity, specify a game server group ARN and the type of
activity to be resumed.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gsg-intro.html}{GameLift FleetIQ Guide}

\strong{Related operations}
\itemize{
\item CreateGameServerGroup
\item ListGameServerGroups
\item DescribeGameServerGroup
\item UpdateGameServerGroup
\item DeleteGameServerGroup
\item ResumeGameServerGroup
\item SuspendGameServerGroup
}
}
\section{Request syntax}{
\preformatted{svc$resume_game_server_group(
  GameServerGroupName = "string",
  ResumeActions = list(
    "REPLACE_INSTANCE_TYPES"
  )
)
}
}

\keyword{internal}
