% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_start_match_backfill}
\alias{gamelift_start_match_backfill}
\title{Finds new players to fill open slots in an existing game session}
\usage{
gamelift_start_match_backfill(TicketId, ConfigurationName,
  GameSessionArn, Players)
}
\arguments{
\item{TicketId}{A unique identifier for a matchmaking ticket. If no ticket ID is
specified here, Amazon GameLift will generate one in the form of a UUID.
Use this identifier to track the match backfill ticket status and
retrieve match results.}

\item{ConfigurationName}{[required] Name of the matchmaker to use for this request. You can use either the
configuration name or ARN value. The ARN of the matchmaker that was used
with the original game session is listed in the GameSession object,
\code{MatchmakerData} property.}

\item{GameSessionArn}{Amazon Resource Name
(\href{https://docs.aws.amazon.com/general/latest/gr/aws-arns-and-namespaces.html}{ARN})
that is assigned to a game session and uniquely identifies it. This is
the same as the game session ID.}

\item{Players}{[required] Match information on all players that are currently assigned to the game
session. This information is used by the matchmaker to find new players
and add them to the existing game.
\itemize{
\item PlayerID, PlayerAttributes, Team -\\\\- This information is
maintained in the GameSession object, \code{MatchmakerData} property, for
all players who are currently assigned to the game session. The
matchmaker data is in JSON syntax, formatted as a string. For more
details, see \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-server.html#match-server-data}{Match Data}.
\item LatencyInMs -\\\\- If the matchmaker uses player latency, include a
latency value, in milliseconds, for the Region that the game session
is currently in. Do not include latency values for any other Region.
}}
}
\value{
A list with the following syntax:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(
  MatchmakingTicket = list(
    TicketId = "string",
    ConfigurationName = "string",
    ConfigurationArn = "string",
    Status = "CANCELLED"|"COMPLETED"|"FAILED"|"PLACING"|"QUEUED"|"REQUIRES_ACCEPTANCE"|"SEARCHING"|"TIMED_OUT",
    StatusReason = "string",
    StatusMessage = "string",
    StartTime = as.POSIXct(
      "2015-01-01"
    ),
    EndTime = as.POSIXct(
      "2015-01-01"
    ),
    Players = list(
      list(
        PlayerId = "string",
        PlayerAttributes = list(
          list(
            S = "string",
            N = 123.0,
            SL = list(
              "string"
            ),
            SDM = list(
              123.0
            )
          )
        ),
        Team = "string",
        LatencyInMs = list(
          123
        )
      )
    ),
    GameSessionConnectionInfo = list(
      GameSessionArn = "string",
      IpAddress = "string",
      DnsName = "string",
      Port = 123,
      MatchedPlayerSessions = list(
        list(
          PlayerId = "string",
          PlayerSessionId = "string"
        )
      )
    ),
    EstimatedWaitTime = 123
  )
)
}\if{html}{\out{</div>}}
}
\description{
Finds new players to fill open slots in an existing game session. This
operation can be used to add players to matched games that start with
fewer than the maximum number of players or to replace players when they
drop out. By backfilling with the same matchmaker used to create the
original match, you ensure that new players meet the match criteria and
maintain a consistent experience throughout the game session. You can
backfill a match anytime after a game session has been created.

To request a match backfill, specify a unique ticket ID, the existing
game session's ARN, a matchmaking configuration, and a set of data that
describes all current players in the game session. If successful, a
match backfill ticket is created and returned with status set to QUEUED.
The ticket is placed in the matchmaker's ticket pool and processed.
Track the status of the ticket to respond as needed.

The process of finding backfill matches is essentially identical to the
initial matchmaking process. The matchmaker searches the pool and groups
tickets together to form potential matches, allowing only one backfill
ticket per potential match. Once the a match is formed, the matchmaker
creates player sessions for the new players. All tickets in the match
are updated with the game session's connection information, and the
GameSession object is updated to include matchmaker data on the new
players. For more detail on how match backfill requests are processed,
see \href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/gamelift-match.html}{How Amazon GameLift FlexMatch Works}.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/match-backfill.html}{Backfill Existing Games with FlexMatch}

\href{https://docs.aws.amazon.com/gamelift/latest/flexmatchguide/gamelift-match.html}{How GameLift FlexMatch Works}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_start_matchmaking]{start_matchmaking}}
\item \code{\link[=gamelift_describe_matchmaking]{describe_matchmaking}}
\item \code{\link[=gamelift_stop_matchmaking]{stop_matchmaking}}
\item \code{\link[=gamelift_accept_match]{accept_match}}
\item \code{\link[=gamelift_start_match_backfill]{start_match_backfill}}
}
}
\section{Request syntax}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{svc$start_match_backfill(
  TicketId = "string",
  ConfigurationName = "string",
  GameSessionArn = "string",
  Players = list(
    list(
      PlayerId = "string",
      PlayerAttributes = list(
        list(
          S = "string",
          N = 123.0,
          SL = list(
            "string"
          ),
          SDM = list(
            123.0
          )
        )
      ),
      Team = "string",
      LatencyInMs = list(
        123
      )
    )
  )
)
}\if{html}{\out{</div>}}
}

\keyword{internal}
