% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_build}
\alias{gamelift_update_build}
\title{Updates metadata in a build resource, including the build name and
version}
\usage{
gamelift_update_build(BuildId, Name, Version)
}
\arguments{
\item{BuildId}{[required] A unique identifier for a build to update. You can use either the build
ID or ARN value.}

\item{Name}{A descriptive label that is associated with a build. Build names do not
need to be unique.}

\item{Version}{Version information that is associated with a build or script. Version
strings do not need to be unique.}
}
\description{
Updates metadata in a build resource, including the build name and
version. To update the metadata, specify the build ID to update and
provide the new values. If successful, a build object containing the
updated metadata is returned.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/gamelift-build-intro.html}{Upload a Custom Server Build}

\strong{Related operations}
\itemize{
\item CreateBuild
\item ListBuilds
\item DescribeBuild
\item UpdateBuild
\item DeleteBuild
}
}
\section{Request syntax}{
\preformatted{svc$update_build(
  BuildId = "string",
  Name = "string",
  Version = "string"
)
}
}

\keyword{internal}
