% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_update_fleet_capacity}
\alias{gamelift_update_fleet_capacity}
\title{Updates capacity settings for a fleet}
\usage{
gamelift_update_fleet_capacity(FleetId, DesiredInstances, MinSize,
  MaxSize)
}
\arguments{
\item{FleetId}{[required] A unique identifier for a fleet to update capacity for. You can use
either the fleet ID or ARN value.}

\item{DesiredInstances}{Number of EC2 instances you want this fleet to host.}

\item{MinSize}{The minimum value allowed for the fleet's instance count. Default if not
set is 0.}

\item{MaxSize}{The maximum value allowed for the fleet's instance count. Default if not
set is 1.}
}
\description{
Updates capacity settings for a fleet. Use this operation to specify the
number of EC2 instances (hosts) that you want this fleet to contain.
Before calling this operation, you may want to call
\code{\link[=gamelift_describe_ec2_instance_limits]{describe_ec2_instance_limits}}
to get the maximum capacity based on the fleet's EC2 instance type.

Specify minimum and maximum number of instances. Amazon GameLift will
not change fleet capacity to values fall outside of this range. This is
particularly important when using auto-scaling (see
\code{\link[=gamelift_put_scaling_policy]{put_scaling_policy}}) to allow capacity
to adjust based on player demand while imposing limits on automatic
adjustments.

To update fleet capacity, specify the fleet ID and the number of
instances you want the fleet to host. If successful, Amazon GameLift
starts or terminates instances so that the fleet's active instance count
matches the desired instance count. You can view a fleet's current
capacity information by calling
\code{\link[=gamelift_describe_fleet_capacity]{describe_fleet_capacity}}. If the
desired instance count is higher than the instance type's limit, the
"Limit Exceeded" exception occurs.

\strong{Learn more}

\href{https://docs.aws.amazon.com/gamelift/latest/developerguide/fleets-intro.html}{Setting up GameLift Fleets}

\strong{Related operations}
\itemize{
\item \code{\link[=gamelift_create_fleet]{create_fleet}}
\item \code{\link[=gamelift_list_fleets]{list_fleets}}
\item \code{\link[=gamelift_delete_fleet]{delete_fleet}}
\item \code{\link[=gamelift_describe_fleet_attributes]{describe_fleet_attributes}}
\item Update fleets:
\itemize{
\item \code{\link[=gamelift_update_fleet_attributes]{update_fleet_attributes}}
\item \code{\link[=gamelift_update_fleet_capacity]{update_fleet_capacity}}
\item \code{\link[=gamelift_update_fleet_port_settings]{update_fleet_port_settings}}
\item \code{\link[=gamelift_update_runtime_configuration]{update_runtime_configuration}}
}
\item \code{\link[=gamelift_start_fleet_actions]{start_fleet_actions}} or
\code{\link[=gamelift_stop_fleet_actions]{stop_fleet_actions}}
}
}
\section{Request syntax}{
\preformatted{svc$update_fleet_capacity(
  FleetId = "string",
  DesiredInstances = 123,
  MinSize = 123,
  MaxSize = 123
)
}
}

\keyword{internal}
